@@ -1,6 +1,5 @@
 buildscript {
 	repositories {
-		maven { url "https://maven.fabric.io/public" }
 		maven { url "https://plugins.gradle.org/m2/" }
 		mavenCentral()
 		jcenter()
@@ -9,8 +8,7 @@ buildscript {
 	dependencies {
 		classpath "com.android.tools.build:gradle:3.4.1"
 		classpath 'com.google.gms:google-services:4.2.0'
-		classpath 'io.fabric.tools:gradle:1.25.4'
-		classpath "im.getsocial:plugin:0.0.17"
+		classpath "im.getsocial:plugin-v7:1.0.10"
 	}
 }
 
@@ -21,7 +19,6 @@ repositories {
 }
 
 apply plugin: 'com.android.application'
-apply plugin: 'io.fabric'
 apply plugin: 'im.getsocial'
 
 getsocial {
@@ -42,9 +39,9 @@ dependencies {
 	implementation 'com.jakewharton:butterknife:10.0.0'
 	implementation 'com.google.code.findbugs:jsr305:3.0.2'
 	implementation 'com.squareup.picasso:picasso:2.5.2'
-	implementation 'com.android.billingclient:billing:1.1'
+	implementation 'com.android.billingclient:billing:3.0.0'
 	implementation "com.google.android.material:material:1.0.0"
-	
+
 	//  GetSocial supports both FCM and GCM. 
 	//  Uncomment following two dependencies and comment out Firebase one to try
 	//	GetSocial push notifications via GCM. Check documentation for more details.
@@ -52,9 +49,8 @@ dependencies {
 	// implementation "com.google.android.gms:play-services-gcm:11.8.0"
 	implementation "com.google.firebase:firebase-messaging:19.0.1"
 
-	implementation('com.crashlytics.sdk.android:crashlytics:2.7.1@aar') {
-		transitive = true
-	}
+	implementation "com.microsoft.appcenter:appcenter-analytics:2.4.0"
+	implementation "com.microsoft.appcenter:appcenter-crashes:2.4.0"
 
 	//adjust partnership
 	implementation 'com.adjust.sdk:adjust-android:4.12.0'
@@ -67,11 +63,11 @@ dependencies {
 }
 
 android {
-	compileSdkVersion 28
+	compileSdkVersion 31
 
 	defaultConfig {
 		minSdkVersion 16
-		targetSdkVersion 28
+		targetSdkVersion 31
 
 		applicationId 'im.getsocial.demo'
 		versionCode 1
@@ -84,12 +80,6 @@ android {
 	productFlavors {
 		publicApp {}
 	}
-	buildTypes {
-		release {
-			ext.enableCrashlytics = false
-		}
-	}
-
 	signingConfigs {
 		debug {
 			storeFile file("keystores/debug.keystore")
