@@ -1,7 +1,6 @@
 package im.getsocial.demo.fragment;
 
 import android.os.Bundle;
-import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -9,12 +8,11 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Spinner;
+import androidx.annotation.Nullable;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import im.getsocial.demo.R;
-import im.getsocial.sdk.CompletionCallback;
-import im.getsocial.sdk.GetSocial;
-import im.getsocial.sdk.GetSocialException;
+import im.getsocial.sdk.Analytics;
 import im.getsocial.sdk.iap.PurchaseData;
 
 import java.util.UUID;
@@ -34,29 +32,29 @@ public class ManualIAPFragment extends BaseFragment {
 	}
 
 	@Override
-	public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+	public View onCreateView(final LayoutInflater inflater, @Nullable final ViewGroup container, @Nullable final Bundle savedInstanceState) {
 		return inflater.inflate(R.layout.fragment_manual_iap, container, false);
 	}
 
 	@Override
-	public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+	public void onViewCreated(final View view, @Nullable final Bundle savedInstanceState) {
 		super.onViewCreated(view, savedInstanceState);
 		_viewContainer = new ViewContainer(view);
 	}
 
 	private void trackPurchase() {
 		final String productType = _viewContainer._productTypeInput.getSelectedItem().toString();
-		final PurchaseData purchaseData = new PurchaseData.Builder()
-				.withProductId(_viewContainer._productIdInput.getText().toString())
-				.withProductTitle(_viewContainer._productTitleInput.getText().toString())
-				.withProductType(productType.equalsIgnoreCase("item") ? PurchaseData.ProductType.ITEM : PurchaseData.ProductType.SUBSCRIPTION)
-				.withPrice(Float.parseFloat(_viewContainer._priceInput.getText().toString()))
-				.withPriceCurrency(_viewContainer._priceCurrencyInput.getText().toString())
-				.withPurchaseId(UUID.randomUUID().toString())
-				.withPurchaseDate(System.currentTimeMillis() / 1000)
-				.build();
+		final PurchaseData purchaseData = new PurchaseData()
+						.withProductId(_viewContainer._productIdInput.getText().toString())
+						.withProductTitle(_viewContainer._productTitleInput.getText().toString())
+						.withProductType(productType.equalsIgnoreCase("item") ? PurchaseData.ProductType.ITEM : PurchaseData.ProductType.SUBSCRIPTION)
+						.withPrice(Float.parseFloat(_viewContainer._priceInput.getText().toString()))
+						.withPriceCurrency(_viewContainer._priceCurrencyInput.getText().toString())
+						.withPurchaseId(UUID.randomUUID().toString())
+						.withPurchaseDate(System.currentTimeMillis() / 1000);
 
-		if (GetSocial.trackPurchaseEvent(purchaseData)) {
+
+		if (Analytics.trackPurchaseEvent(purchaseData)) {
 			_log.logInfoAndToast("Successfully tracked purchase data");
 		} else {
 			_log.logErrorAndToast("Failed to track purchase data.");
@@ -83,16 +81,11 @@ public class ManualIAPFragment extends BaseFragment {
 		@BindView(R.id.button_send_purchase)
 		Button _sendPurchase;
 
-		ViewContainer(View view) {
+		ViewContainer(final View view) {
 			ButterKnife.bind(this, view);
-			_sendPurchase.setOnClickListener(new View.OnClickListener() {
-				@Override
-				public void onClick(View view) {
-					trackPurchase();
-				}
-			});
-			String[] productTypes = new String[] { "item", "subscription" };
-			ArrayAdapter<String> adapter = new ArrayAdapter<String>(getContext(), android.R.layout.simple_spinner_dropdown_item, productTypes);
+			_sendPurchase.setOnClickListener(view1 -> trackPurchase());
+			final String[] productTypes = new String[] {"item", "subscription"};
+			final ArrayAdapter<String> adapter = new ArrayAdapter<>(getContext(), android.R.layout.simple_spinner_dropdown_item, productTypes);
 			_productTypeInput.setAdapter(adapter);
 		}
 
