@@ -1,18 +1,18 @@
 /*
-*    	Copyright 2015-2017 GetSocial B.V.
-*
-*	Licensed under the Apache License, Version 2.0 (the "License");
-*	you may not use this file except in compliance with the License.
-*	You may obtain a copy of the License at
-*
-*    	http://www.apache.org/licenses/LICENSE-2.0
-*
-*	Unless required by applicable law or agreed to in writing, software
-*	distributed under the License is distributed on an "AS IS" BASIS,
-*	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*	See the License for the specific language governing permissions and
-*	limitations under the License.
-*/
+ *    	Copyright 2015-2017 GetSocial B.V.
+ *
+ *	Licensed under the Apache License, Version 2.0 (the "License");
+ *	you may not use this file except in compliance with the License.
+ *	You may obtain a copy of the License at
+ *
+ *    	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *	Unless required by applicable law or agreed to in writing, software
+ *	distributed under the License is distributed on an "AS IS" BASIS,
+ *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *	See the License for the specific language governing permissions and
+ *	limitations under the License.
+ */
 
 package im.getsocial.demo.fragment;
 
@@ -20,19 +20,18 @@ import android.graphics.Bitmap;
 import android.graphics.drawable.BitmapDrawable;
 import android.net.Uri;
 import android.os.Bundle;
-import androidx.annotation.Nullable;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.Spinner;
 import android.widget.Toast;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnCheckedChanged;
@@ -41,26 +40,31 @@ import com.squareup.picasso.MemoryPolicy;
 import com.squareup.picasso.Picasso;
 import im.getsocial.demo.R;
 import im.getsocial.demo.ui.PickActionView;
+import im.getsocial.demo.utils.DynamicUi;
 import im.getsocial.demo.utils.VideoUtils;
 import im.getsocial.sdk.Callback;
-import im.getsocial.sdk.ErrorCode;
-import im.getsocial.sdk.GetSocial;
-import im.getsocial.sdk.GetSocialException;
+import im.getsocial.sdk.Communities;
+import im.getsocial.sdk.FailureCallback;
+import im.getsocial.sdk.GetSocialError;
 import im.getsocial.sdk.actions.Action;
-import im.getsocial.sdk.actions.ActionListener;
-import im.getsocial.sdk.activities.ActivityPost;
-import im.getsocial.sdk.activities.ActivityPostContent;
+import im.getsocial.sdk.communities.ActivitiesQuery;
+import im.getsocial.sdk.communities.ActivityButton;
+import im.getsocial.sdk.communities.ActivityContent;
+import im.getsocial.sdk.communities.GetSocialActivity;
+import im.getsocial.sdk.communities.PostActivityTarget;
+import im.getsocial.sdk.communities.UserId;
 import im.getsocial.sdk.media.MediaAttachment;
-import im.getsocial.sdk.ui.GetSocialUi;
-import im.getsocial.sdk.ui.UiAction;
-import im.getsocial.sdk.ui.activities.ActionButtonListener;
-import im.getsocial.sdk.ui.activities.ActivityFeedViewBuilder;
+import im.getsocial.sdk.ui.communities.ActivityDetailsViewBuilder;
+import im.getsocial.sdk.ui.communities.ActivityFeedViewBuilder;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
 import static com.squareup.picasso.Picasso.with;
 
-public class PostActivityFragment extends BaseFragment implements Callback<ActivityPost> {
+public class PostActivityFragment extends BaseFragment implements Callback<GetSocialActivity>, FailureCallback {
 
 	private static final int MAX_WIDTH = 1024;
 	private static final int REQUEST_PICK_CUSTOM_IMAGE = 0x1;
@@ -69,16 +73,67 @@ public class PostActivityFragment extends BaseFragment implements Callback<Activ
 	private ViewContainer _viewContainer;
 	private VideoUtils.VideoDescriptor _videoDescriptor;
 	private Bitmap _originalImage;
+	@Nullable
+	private String _activityId;
+
+	@Nullable
+	private String _postTopic;
+
+	@Nullable
+	private String _postGroup;
+
+	@Nullable
+	private MediaAttachment _videoAttachment;
+	@Nullable
+	private MediaAttachment _imageAttachment;
+
+	final List<DynamicUi.DynamicInputHolder> _propertiesHolder = new ArrayList<>();
+
+	public static Fragment updateActivity(final String activityId) {
+		final PostActivityFragment fragment = new PostActivityFragment();
+		final Bundle args = new Bundle();
+		args.putString("activity", activityId);
+		fragment.setArguments(args);
+		return fragment;
+	}
+
+	public static Fragment postToTopic(final String topicId) {
+		final PostActivityFragment fragment = new PostActivityFragment();
+		final Bundle args = new Bundle();
+		args.putString("topic", topicId);
+		fragment.setArguments(args);
+		return fragment;
+	}
+
+	public static Fragment postToGroup(final String id) {
+		final PostActivityFragment fragment = new PostActivityFragment();
+		final Bundle args = new Bundle();
+		args.putString("group", id);
+		fragment.setArguments(args);
+		return fragment;
+	}
 
 	@Override
-	public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+	public View onCreateView(final LayoutInflater inflater, @Nullable final ViewGroup container, @Nullable final Bundle savedInstanceState) {
 		return inflater.inflate(R.layout.fragment_post_activity, container, false);
 	}
 
 	@Override
-	public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+	public void onViewCreated(final View view, @Nullable final Bundle savedInstanceState) {
 		super.onViewCreated(view, savedInstanceState);
 		_viewContainer = new ViewContainer(view);
+		final Bundle bundle = getArguments();
+		if (bundle != null) {
+			_activityId = bundle.getString("activity");
+
+			if (_activityId != null) {
+				Communities.getActivity(_activityId, post -> {
+					_viewContainer.setPost(post);
+				}, this::onFailure);
+			}
+			_postTopic = bundle.getString("topic");
+			_postGroup = bundle.getString("group");
+		}
 	}
 
 	@Override
@@ -98,15 +153,14 @@ public class PostActivityFragment extends BaseFragment implements Callback<Activ
 		final String text = _viewContainer._postText.getText().toString();
 
 		final String buttonTitle = _viewContainer._buttonTitle.getText().toString();
-		final String oldButtonAction = _viewContainer._buttonAction.getText().toString();
 		final Action action = _viewContainer._pickActionView.getAction();
 
-		boolean hasText = !TextUtils.isEmpty(text);
-		boolean hasImage = _viewContainer._hasImage.isChecked() && _originalImage != null;
-		boolean hasVideo = _viewContainer._hasVideo.isChecked() && _videoDescriptor != null;
-		boolean hasButton = _viewContainer._hasButton.isChecked()
-				&& !buttonTitle.isEmpty()
-				&& (!oldButtonAction.isEmpty() || action != null);
+		final boolean hasText = !TextUtils.isEmpty(text);
+		final boolean hasImage = _viewContainer._hasImage.isChecked() && _originalImage != null || _imageAttachment != null;
+		final boolean hasVideo = _viewContainer._hasVideo.isChecked() && _videoDescriptor != null || _videoAttachment != null;
+		final boolean hasButton = _viewContainer._hasButton.isChecked()
+						&& !buttonTitle.isEmpty()
+						&& (action != null);
 
 		if (!hasText && !hasButton && !hasImage && !hasVideo) {
 			hideLoadingAndShowError("Can not post activity without any data");
@@ -114,113 +168,107 @@ public class PostActivityFragment extends BaseFragment implements Callback<Activ
 			return;
 		}
 
-		boolean postToGlobalFeed = _viewContainer._feed.getSelectedItemPosition() == 0;
-		if (postToGlobalFeed && GetSocial.User.isAnonymous()) {
-			_viewContainer._image.setEnabled(true);
-			hideLoading();
-			showAuthorizeUserDialogForPendingAction("Post to global feed", new UiAction.Pending() {
-				@Override
-				public void proceed() {
-					doPost();
-				}
-			});
-			return;
-		}
-
-		ActivityPostContent.Builder builder = new ActivityPostContent.Builder();
+		final ActivityContent builder = new ActivityContent();
 		if (hasText) {
 			builder.withText(text);
 		}
 		if (hasButton) {
-			if (!oldButtonAction.isEmpty()) {
-				builder.withButton(buttonTitle, oldButtonAction);
-			}
-			if (action != null) {
-				builder.withButton(buttonTitle, action);
-			}
+			builder.withButton(ActivityButton.create(buttonTitle, action));
 		}
 		if (hasImage) {
-			builder.withMediaAttachment(MediaAttachment.image(_originalImage));
+			builder.addAttachment(_imageAttachment != null ? _imageAttachment : MediaAttachment.image(_originalImage));
 		}
 		if (hasVideo) {
-			builder.withMediaAttachment(MediaAttachment.video(_videoDescriptor._video));
+			builder.addAttachment(_videoAttachment != null ? _videoAttachment : MediaAttachment.video(_videoDescriptor._video));
+		}
+		for (final DynamicUi.DynamicInputHolder property : _propertiesHolder) {
+			builder.addProperty(property.getText(0), property.getText(1));
+		}
+
+		if (_activityId != null) {
+			showLoading("Updating activity", "Wait...");
+			Communities.updateActivity(_activityId, builder, this, this);
+			return;
 		}
 		showLoading("Posting activity", "Wait...");
+		Communities.postActivity(builder, target(), this, this);
+	}
 
-		if (postToGlobalFeed) {
-			GetSocial.postActivityToGlobalFeed(builder.build(), this);
-		} else {
-			GetSocial.postActivityToFeed(ActivitiesFragment.CUSTOM_FEED_NAME, builder.build(), this);
+	private PostActivityTarget target() {
+		if (_postTopic != null) {
+			return PostActivityTarget.topic(_postTopic);
 		}
+		if (_postGroup != null) {
+			return PostActivityTarget.group(_postGroup);
+		}
+		return PostActivityTarget.timeline();
+	}
+
+	private ActivitiesQuery query() {
+		if (_postTopic != null) {
+			return ActivitiesQuery.activitiesInTopic(_postTopic);
+		}
+		if (_postGroup != null) {
+			return ActivitiesQuery.activitiesInGroup(_postGroup);
+		}
+		return ActivitiesQuery.feedOf(UserId.currentUser());
 	}
 
 	@Override
-	public void onSuccess(ActivityPost activityPost) {
+	public void onSuccess(final GetSocialActivity activity) {
 		hideLoading();
 		_viewContainer._image.setEnabled(true);
 
 		Toast.makeText(getContext(), "Activity was successfully posted", Toast.LENGTH_SHORT).show();
 
-		boolean postToGlobalFeed = _viewContainer._feed.getSelectedItemPosition() == 0;
-		ActivityFeedViewBuilder feedView = postToGlobalFeed ? GetSocialUi.createGlobalActivityFeedView() : GetSocialUi.createActivityFeedView(ActivitiesFragment.CUSTOM_FEED_NAME);
+		if (_activityId != null) {
+			ActivityDetailsViewBuilder.create(_activityId)
+							.setActionListener(_activityListener.dependencies().actionListener())
+							.setShowActivityFeedView(true).show();
+			return;
+		}
 
-		feedView.setButtonActionListener(new ActionButtonListener() {
-			@Override
-			public void onButtonClicked(String action, ActivityPost post) {
-				Toast.makeText(getContext(), "Activity Feed button clicked, action: " + action, Toast.LENGTH_SHORT).show();
-			}
-		}).setActionListener(new ActionListener() {
-			@Override
-			public boolean handleAction(Action action) {
-				return _activityListener.dependencies().actionListener().handleAction(action);
-			}
-		}).show();
+		ActivityFeedViewBuilder.create(query())
+						.setActionListener(_activityListener.dependencies().actionListener())
+						.show();
 	}
 
 	@Override
-	public void onFailure(GetSocialException exception) {
+	public void onFailure(final GetSocialError exception) {
 		_viewContainer._image.setEnabled(true);
-		if (exception.getErrorCode() == ErrorCode.ACTIVITY_REJECTED) {
-			hideLoading();
-			Toast.makeText(getContext(), "Your activity was rejected during the moderation: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
-		} else {
-			hideLoadingAndShowError(exception.getMessage());
-		}
+		hideLoadingAndShowError(exception.getMessage());
 	}
 
-	private void hideLoadingAndShowError(String message) {
+	private void hideLoadingAndShowError(final String message) {
 		hideLoading();
 		Toast.makeText(getContext(), "Failed to post activity: " + message, Toast.LENGTH_SHORT).show();
 	}
 
 	private void loadOriginalImage(final Uri imageUri) {
-		new Thread(new Runnable() {
-			@Override
-			public void run() {
-				try {
-					_originalImage = Picasso.with(getContext()).load(imageUri).get();
-				} catch (IOException e) {
-					_viewContainer._image = null;
-					Toast.makeText(getContext(), "Could not load original image: " + e.getMessage(), Toast.LENGTH_SHORT).show();
-				}
+		new Thread(() -> {
+			try {
+				_originalImage = Picasso.with(getContext()).load(imageUri).get();
+			} catch (final IOException e) {
+				_viewContainer._image = null;
+				Toast.makeText(getContext(), "Could not load original image: " + e.getMessage(), Toast.LENGTH_SHORT).show();
 			}
 		}).start();
 	}
 
 	@Override
-	protected void onImagePickedFromDevice(Uri imageUri, int requestCode) {
+	protected void onImagePickedFromDevice(final Uri imageUri, final int requestCode) {
 		if (requestCode == REQUEST_PICK_CUSTOM_IMAGE) {
 			with(getContext())
-					.load(imageUri)
-					.resize(MAX_WIDTH, 0)
-					.memoryPolicy(MemoryPolicy.NO_CACHE)
-					.into(_viewContainer._image);
+							.load(imageUri)
+							.resize(MAX_WIDTH, 0)
+							.memoryPolicy(MemoryPolicy.NO_CACHE)
+							.into(_viewContainer._image);
 			loadOriginalImage(imageUri);
 		}
 	}
 
 	@Override
-	protected void onVideoPickedFromDevice(Uri videoUri, int requestCode) {
+	protected void onVideoPickedFromDevice(final Uri videoUri, final int requestCode) {
 		if (requestCode == REQUEST_PICK_CUSTOM_VIDEO) {
 			_videoDescriptor = VideoUtils.open(getContext(), videoUri);
 			if (_videoDescriptor == null) {
@@ -244,8 +292,6 @@ public class PostActivityFragment extends BaseFragment implements Callback<Activ
 		LinearLayout _buttonDataContainer;
 		@BindView(R.id.input_button_title)
 		EditText _buttonTitle;
-		@BindView(R.id.input_button_action)
-		EditText _buttonAction;
 		@BindView(R.id.checkbox_has_image)
 		CheckBox _hasImage;
 		@BindView(R.id.checkbox_has_video)
@@ -254,16 +300,21 @@ public class PostActivityFragment extends BaseFragment implements Callback<Activ
 		ImageView _image;
 		@BindView(R.id.image_view_post_video)
 		ImageView _video;
-		@BindView(R.id.selector_feed)
-		Spinner _feed;
+		@BindView(R.id.properties)
+		LinearLayout _properties;
 		@BindView(R.id.select_video)
 		Button _selectVideo;
 
-		ViewContainer(View view) {
+		ViewContainer(final View view) {
 			ButterKnife.bind(this, view);
+		}
 
-			_feed.setAdapter(ArrayAdapter.createFromResource(getContext(),
-					R.array.activityFeedsItems, android.R.layout.simple_spinner_item));
+		private void createRow(final String k, final String v) {
+			final DynamicUi.DynamicInputHolder inputHolder = DynamicUi.createDynamicTextRow(getContext(), _properties, _propertiesHolder, "Key", "Value");
+			final EditText key = inputHolder.getView(0);
+			final EditText val = inputHolder.getView(1);
+			key.setText(k);
+			val.setText(v);
 		}
 
 		@OnClick(R.id.button_post)
@@ -282,31 +333,60 @@ public class PostActivityFragment extends BaseFragment implements Callback<Activ
 		}
 
 		@OnCheckedChanged(R.id.checkbox_has_button)
-		void toggleHasButton(boolean hasButton) {
+		void toggleHasButton(final boolean hasButton) {
 			_buttonDataContainer.setVisibility(hasButton ? View.VISIBLE : View.GONE);
 		}
 
 		@OnCheckedChanged(R.id.checkbox_has_image)
-		void toggleHasImage(boolean hasImage) {
-			if (hasImage) {
-				_video = null;
-				_hasVideo.setChecked(false);
+		void toggleHasImage(final boolean hasImage) {
+			if (!hasImage) {
+				_imageAttachment = null;
 			}
 			_image.setVisibility(hasImage ? View.VISIBLE : View.GONE);
-			_originalImage = hasImage ? ((BitmapDrawable)_image.getDrawable()).getBitmap() : null;
+			_originalImage = hasImage ? ((BitmapDrawable) _image.getDrawable()).getBitmap() : null;
 		}
 
 		@OnCheckedChanged(R.id.checkbox_has_video)
-		void toggleHasVideo(boolean hasVideo) {
-			if (hasVideo) {
-				_hasImage.setChecked(false);
-			} else {
+		void toggleHasVideo(final boolean hasVideo) {
+			if (!hasVideo) {
 				_video.setImageDrawable(null);
 				_videoDescriptor = null;
 				_video.setVisibility(View.GONE);
+				_videoAttachment = null;
 			}
 			_selectVideo.setVisibility(hasVideo ? View.VISIBLE : View.GONE);
 		}
 
+		@OnClick(R.id.add_property)
+		void addProperty() {
+			createRow("", "");
+		}
+
+		public void setPost(final GetSocialActivity post) {
+			_postText.setText(post.getText());
+			final ActivityButton button = post.getButton();
+			if (button != null) {
+				_hasButton.setChecked(true);
+				_buttonTitle.setText(button.getTitle());
+				_pickActionView.setAction(button.getAction());
+			} else {
+				_hasButton.setChecked(false);
+			}
+			for (final MediaAttachment attachment : post.getAttachments()) {
+				if (attachment.getVideoUrl() != null) {
+					_hasVideo.setChecked(true);
+					Picasso.with(getContext()).load(attachment.getImageUrl()).into(_video);
+					_video.setVisibility(View.VISIBLE);
+					_videoAttachment = attachment;
+				} else if (attachment.getImageUrl() != null) {
+					_hasImage.setChecked(true);
+					Picasso.with(getContext()).load(attachment.getImageUrl()).into(_image);
+					_imageAttachment = attachment;
+				}
+			}
+			for (final Map.Entry<String, String> property : post.getProperties().entrySet()) {
+				createRow(property.getKey(), property.getValue());
+			}
+		}
 	}
 }
