@@ -4,38 +4,63 @@ import android.app.AlertDialog;
 import android.content.ClipData;
 import android.content.ClipboardManager;
 import android.content.Context;
-import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 import im.getsocial.demo.adapter.MenuItem;
-import im.getsocial.sdk.Callback;
-import im.getsocial.sdk.CompletionCallback;
 import im.getsocial.sdk.GetSocial;
-import im.getsocial.sdk.GetSocialException;
+import im.getsocial.sdk.PromoCodes;
+import im.getsocial.sdk.communities.CurrentUser;
+import im.getsocial.sdk.communities.UserUpdate;
 import im.getsocial.sdk.invites.InviteContent;
 import im.getsocial.sdk.invites.InviteTextPlaceholders;
 import im.getsocial.sdk.invites.LinkParams;
 import im.getsocial.sdk.promocodes.PromoCode;
-import im.getsocial.sdk.promocodes.PromoCodeBuilder;
-import im.getsocial.sdk.ui.GetSocialUi;
+import im.getsocial.sdk.promocodes.PromoCodeContent;
+import im.getsocial.sdk.ui.invites.InvitesViewBuilder;
 
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 
 public class PromoCodesFragment extends BaseListFragment {
 
 	private static final String PROMO_CODE_PROPERTY = "my_promo_code";
 
+	private static void shareCode(final String promoCode) {
+		final InviteContent content = new InviteContent();
+		content.setText("Use my Promo Code to get a personal discount: " + InviteTextPlaceholders.PLACEHOLDER_PROMO_CODE + " . " + InviteTextPlaceholders.PLACEHOLDER_APP_INVITE_URL);
+		content.setLinkParams(new HashMap<String, Object>() {
+			{
+				put(LinkParams.KEY_PROMO_CODE, promoCode);
+			}
+		});
+		InvitesViewBuilder.create().setCustomInviteContent(content).show();
+	}
+
+	private static void copyCode(final Context context, final String promoCode) {
+		final ClipboardManager clipboard = (ClipboardManager) context.getSystemService(Context.CLIPBOARD_SERVICE);
+		final ClipData clip = ClipData.newPlainText("GetSocial Promo Code", promoCode);
+		clipboard.setPrimaryClip(clip);
+	}
+
+	private static String formatInfo(final PromoCode promoCode) {
+		return String.format("code=%s\ndata=%s\nmaxClaim=%d\nclaimCount=%d\nstartDate=%s" +
+										"\nendDate=%s\nenabled=%b\nclaimable=%b\ncreator=%s",
+						promoCode.getCode(), promoCode.getProperties(), promoCode.getMaxClaimCount(),
+						promoCode.getClaimCount(), promoCode.getStartDate(), promoCode.getEndDate(),
+						promoCode.isEnabled(), promoCode.isClaimable(), promoCode.getCreator().getDisplayName());
+	}
+
 	@Override
 	protected List<MenuItem> createListData() {
 		return Arrays.asList(
-				MenuItem.builder("My Promo Code").withAction(new MyPromoCodeAction()).build(),
-				navigationListItem("Create Promo Code", CreatePromoCodeFragment.class),
-				MenuItem.builder("Claim Promo Code").withAction(new ClaimPromoCodeAction()).build(),
-				MenuItem.builder("Promo Code Info").withAction(new InfoPromoCodeAction()).build()
+						MenuItem.builder("My Promo Code").withAction(new MyPromoCodeAction()).build(),
+						navigationListItem("Create Promo Code", CreatePromoCodeFragment.class),
+						MenuItem.builder("Claim Promo Code").withAction(new ClaimPromoCodeAction()).build(),
+						MenuItem.builder("Promo Code Info").withAction(new InfoPromoCodeAction()).build()
 		);
 	}
 
@@ -49,123 +74,38 @@ public class PromoCodesFragment extends BaseListFragment {
 		return "Promo Codes";
 	}
 
-	private class MyPromoCodeAction implements MenuItem.Action {
-
-		@Override
-		public void execute() {
-			if (GetSocial.User.hasPrivateProperty(PROMO_CODE_PROPERTY)) {
-				showPromoCodeDialog(GetSocial.User.getPrivateProperty(PROMO_CODE_PROPERTY));
-			} else {
-				showLoading("Promo Code", "Creating...");
-				GetSocial.createPromoCode(PromoCodeBuilder.createRandomCode().addData("my_promo_code", "true"), new Callback<PromoCode>() {
-					@Override
-					public void onSuccess(final PromoCode result) {
-						GetSocial.User.setPrivateProperty(PROMO_CODE_PROPERTY, result.getCode(), new CompletionCallback() {
-							@Override
-							public void onSuccess() {
-								hideLoading();
-								showPromoCodeDialog(result.getCode());
-							}
-
-							@Override
-							public void onFailure(GetSocialException exception) {
-								hideLoading();
-								Toast.makeText(getContext(), "Failed to save a Promo Code: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
-							}
-						});
-					}
-
-					@Override
-					public void onFailure(GetSocialException exception) {
-						hideLoading();
-						Toast.makeText(getContext(), "Failed to create a Promo Code: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
-					}
-				});
-			}
-		}
-	}
-
-	private class ClaimPromoCodeAction implements MenuItem.Action {
-
-		@Override
-		public void execute() {
-			showDialogWithEditText(
-					"Claim Promo Code",
-					"Promo Code...",
-					new ClaimEditAction(),
-					new DismissEditAction()
-			);
-		}
-	}
-
-	private class InfoPromoCodeAction implements MenuItem.Action {
-
-		@Override
-		public void execute() {
-			showDialogWithEditText(
-					"Promo Code Info",
-					"Promo Code...",
-					new InfoEditAction(),
-					new DismissEditAction()
-			);
-		}
-	}
-
 	private void showPromoCodeDialog(final String promoCode) {
 		showDialog(
-				"Promo Code",
-				promoCode,
-				new ShareAction(promoCode),
-				new CopyAction(promoCode),
-				new InfoAction(promoCode),
-				new DismissAction()
+						"Promo Code",
+						promoCode,
+						new ShareAction(promoCode),
+						new CopyAction(promoCode),
+						new InfoAction(promoCode),
+						new DismissAction()
 		);
 	}
 
 	public void showFullPromoCodeDialog(final PromoCode promoCode) {
 		showDialog(
-				"Promo Code",
-				formatInfo(promoCode),
-				new ShareAction(promoCode.getCode()),
-				new CopyAction(promoCode.getCode()),
-				new DismissAction()
+						"Promo Code",
+						formatInfo(promoCode),
+						new ShareAction(promoCode.getCode()),
+						new CopyAction(promoCode.getCode()),
+						new DismissAction()
 		);
 	}
 
-	private static void shareCode(final String promoCode) {
-		GetSocialUi.createInvitesView()
-				.setCustomInviteContent(InviteContent.createBuilder().withText("Use my Promo Code to get a personal discount: " + InviteTextPlaceholders.PLACEHOLDER_PROMO_CODE + " . " + InviteTextPlaceholders.PLACEHOLDER_APP_INVITE_URL).build())
-				.setLinkParams(new LinkParams() {
-					{ put(LinkParams.KEY_PROMO_CODE, promoCode); }
-				})
-				.show();
-	}
-
-	private static void copyCode(Context context, String promoCode) {
-		ClipboardManager clipboard = (ClipboardManager) context.getSystemService(Context.CLIPBOARD_SERVICE);
-		ClipData clip = ClipData.newPlainText("GetSocial Promo Code", promoCode);
-		clipboard.setPrimaryClip(clip);
-	}
-
-	private void showInfo(PromoCode promoCode) {
+	private void showInfo(final PromoCode promoCode) {
 		new AlertDialog.Builder(getContext())
-				.setTitle("Promo Code")
-				.setMessage(formatInfo(promoCode))
-				.setCancelable(true)
-				.create()
-				.show();
+						.setTitle("Promo Code")
+						.setMessage(formatInfo(promoCode))
+						.setCancelable(true)
+						.create()
+						.show();
 	}
 
-	private static String formatInfo(PromoCode promoCode) {
-		return String.format("code=%s\ndata=%s\nmaxClaim=%d\nclaimCount=%d\nstartDate=%s" +
-						"\nendDate=%s\nenabled=%b\nclaimable=%b\ncreator=%s",
-				promoCode.getCode(), promoCode.getData(), promoCode.getMaxClaimCount(),
-				promoCode.getClaimCount(), promoCode.getStartDate(), promoCode.getEndDate(),
-				promoCode.isEnabled(), promoCode.isClaimable(), promoCode.getCreator().getDisplayName());
-	}
-
-	private void showDialogWithEditText(String title, String hint, EditDialogAction... actions) {
-		LinearLayout linearLayout = new LinearLayout(getContext());
+	private void showDialogWithEditText(final String title, final String hint, final EditDialogAction... actions) {
+		final LinearLayout linearLayout = new LinearLayout(getContext());
 		linearLayout.setPadding(20, 20, 20, 20);
 		linearLayout.setOrientation(LinearLayout.VERTICAL);
 
@@ -176,28 +116,23 @@ public class PromoCodesFragment extends BaseListFragment {
 		linearLayout.addView(inputView);
 
 		final AlertDialog alertDialog = new AlertDialog.Builder(getContext())
-				.setView(linearLayout)
-				.setTitle(title)
-				.setCancelable(true)
-				.create();
+						.setView(linearLayout)
+						.setTitle(title)
+						.setCancelable(true)
+						.create();
 
 		for (final EditDialogAction dialogAction : actions) {
-			Button button = new Button(getContext());
+			final Button button = new Button(getContext());
 			button.setText(dialogAction._name);
-			button.setOnClickListener(new View.OnClickListener() {
-				@Override
-				public void onClick(View view) {
-					dialogAction.onClick(alertDialog, inputView.getText().toString());
-				}
-			});
+			button.setOnClickListener(view -> dialogAction.onClick(alertDialog, inputView.getText().toString()));
 			linearLayout.addView(button);
 		}
 
 		alertDialog.show();
 	}
 
-	private void showDialog(String title, String contentText, DialogAction... actions) {
-		LinearLayout linearLayout = new LinearLayout(getContext());
+	private void showDialog(final String title, final String contentText, final DialogAction... actions) {
+		final LinearLayout linearLayout = new LinearLayout(getContext());
 		linearLayout.setPadding(20, 20, 20, 20);
 		linearLayout.setOrientation(LinearLayout.VERTICAL);
 
@@ -208,20 +143,15 @@ public class PromoCodesFragment extends BaseListFragment {
 		linearLayout.addView(textView);
 
 		final AlertDialog alertDialog = new AlertDialog.Builder(getContext())
-				.setView(linearLayout)
-				.setTitle(title)
-				.setCancelable(true)
-				.create();
+						.setView(linearLayout)
+						.setTitle(title)
+						.setCancelable(true)
+						.create();
 
 		for (final DialogAction dialogAction : actions) {
-			Button button = new Button(getContext());
+			final Button button = new Button(getContext());
 			button.setText(dialogAction._name);
-			button.setOnClickListener(new View.OnClickListener() {
-				@Override
-				public void onClick(View view) {
-					dialogAction.onClick(alertDialog);
-				}
-			});
+			button.setOnClickListener(view -> dialogAction.onClick(alertDialog));
 			linearLayout.addView(button);
 		}
 
@@ -231,11 +161,76 @@ public class PromoCodesFragment extends BaseListFragment {
 	private abstract static class DialogAction {
 		final String _name;
 
-		DialogAction(String name) {
+		DialogAction(final String name) {
 			_name = name;
 		}
 
 		public abstract void onClick(AlertDialog dialog);
+	}
+
+	private abstract static class EditDialogAction {
+		final String _name;
+
+		EditDialogAction(final String name) {
+			_name = name;
+		}
+
+		public abstract void onClick(AlertDialog dialog, String text);
+	}
+
+	private class MyPromoCodeAction implements MenuItem.Action {
+
+		@Override
+		public void execute() {
+			final CurrentUser user = GetSocial.getCurrentUser();
+			if (user == null) {
+				Toast.makeText(getContext(), "SDK is not initialized", Toast.LENGTH_SHORT).show();
+				return;
+			}
+			if (user.getPrivateProperties().containsKey(PROMO_CODE_PROPERTY)) {
+				showPromoCodeDialog(user.getPrivateProperties().get(PROMO_CODE_PROPERTY));
+			} else {
+				showLoading("Promo Code", "Creating...");
+				PromoCodes.create(PromoCodeContent.createRandomCode().addProperty("my_promo_code", "true"), result -> {
+					GetSocial.getCurrentUser().updateDetails(new UserUpdate().setPrivateProperty(PROMO_CODE_PROPERTY, result.getCode()), () -> {
+						hideLoading();
+						showPromoCodeDialog(result.getCode());
+					}, error -> {
+						hideLoading();
+						Toast.makeText(getContext(), "Failed to save a Promo Code: " + error.getMessage(), Toast.LENGTH_SHORT).show();
+					});
+				}, exception -> {
+					hideLoading();
+					Toast.makeText(getContext(), "Failed to create a Promo Code: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
+				});
+			}
+		}
+	}
+
+	private class ClaimPromoCodeAction implements MenuItem.Action {
+
+		@Override
+		public void execute() {
+			showDialogWithEditText(
+							"Claim Promo Code",
+							"Promo Code...",
+							new ClaimEditAction(),
+							new DismissEditAction()
+			);
+		}
+	}
+
+	private class InfoPromoCodeAction implements MenuItem.Action {
+
+		@Override
+		public void execute() {
+			showDialogWithEditText(
+							"Promo Code Info",
+							"Promo Code...",
+							new InfoEditAction(),
+							new DismissEditAction()
+			);
+		}
 	}
 
 	private class DismissAction extends DialogAction {
@@ -245,7 +240,7 @@ public class PromoCodesFragment extends BaseListFragment {
 		}
 
 		@Override
-		public void onClick(AlertDialog dialog) {
+		public void onClick(final AlertDialog dialog) {
 			dialog.dismiss();
 		}
 	}
@@ -253,13 +248,13 @@ public class PromoCodesFragment extends BaseListFragment {
 	private class CopyAction extends DialogAction {
 		private final String _code;
 
-		CopyAction(String code) {
+		CopyAction(final String code) {
 			super("Copy");
 			_code = code;
 		}
 
 		@Override
-		public void onClick(AlertDialog dialog) {
+		public void onClick(final AlertDialog dialog) {
 			copyCode(dialog.getContext(), _code);
 			dialog.dismiss();
 		}
@@ -268,13 +263,13 @@ public class PromoCodesFragment extends BaseListFragment {
 	private class ShareAction extends DialogAction {
 		private final String _code;
 
-		ShareAction(String code) {
+		ShareAction(final String code) {
 			super("Share");
 			_code = code;
 		}
 
 		@Override
-		public void onClick(AlertDialog dialog) {
+		public void onClick(final AlertDialog dialog) {
 			dialog.dismiss();
 			shareCode(_code);
 		}
@@ -283,7 +278,7 @@ public class PromoCodesFragment extends BaseListFragment {
 	private class InfoAction extends DialogAction {
 		private final String _code;
 
-		InfoAction(String promoCode) {
+		InfoAction(final String promoCode) {
 			super("Info");
 			_code = promoCode;
 		}
@@ -291,31 +286,15 @@ public class PromoCodesFragment extends BaseListFragment {
 		@Override
 		public void onClick(final AlertDialog dialog) {
 			showLoading("Promo Code", "Loading...");
-			GetSocial.getPromoCode(_code, new Callback<PromoCode>() {
-				@Override
-				public void onSuccess(PromoCode result) {
-					hideLoading();
-					dialog.dismiss();
-					showFullPromoCodeDialog(result);
-				}
-
-				@Override
-				public void onFailure(GetSocialException exception) {
-					hideLoading();
-					Toast.makeText(getContext(), "Failed to load Promo Code info: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
-				}
+			PromoCodes.get(_code, result -> {
+				hideLoading();
+				dialog.dismiss();
+				showFullPromoCodeDialog(result);
+			}, exception -> {
+				hideLoading();
+				Toast.makeText(getContext(), "Failed to load Promo Code info: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
 			});
 		}
-	}
-
-	private abstract static class EditDialogAction {
-		final String _name;
-
-		EditDialogAction(String name) {
-			_name = name;
-		}
-
-		public abstract void onClick(AlertDialog dialog, String text);
 	}
 
 	private class DismissEditAction extends EditDialogAction {
@@ -325,7 +304,7 @@ public class PromoCodesFragment extends BaseListFragment {
 		}
 
 		@Override
-		public void onClick(AlertDialog dialog, String text) {
+		public void onClick(final AlertDialog dialog, final String text) {
 			dialog.dismiss();
 		}
 	}
@@ -337,18 +316,12 @@ public class PromoCodesFragment extends BaseListFragment {
 		}
 
 		@Override
-		public void onClick(final AlertDialog dialog, String text) {
-			GetSocial.getPromoCode(text, new Callback<PromoCode>() {
-				@Override
-				public void onSuccess(PromoCode result) {
-					dialog.dismiss();
-					showInfo(result);
-				}
-
-				@Override
-				public void onFailure(GetSocialException exception) {
-					Toast.makeText(getContext(), "Failed to get Promo Code: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
-				}
+		public void onClick(final AlertDialog dialog, final String text) {
+			PromoCodes.get(text, result -> {
+				dialog.dismiss();
+				showInfo(result);
+			}, exception -> {
+				Toast.makeText(getContext(), "Failed to get Promo Code: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
 			});
 		}
 	}
@@ -360,19 +333,13 @@ public class PromoCodesFragment extends BaseListFragment {
 		}
 
 		@Override
-		public void onClick(final AlertDialog dialog, String text) {
-			GetSocial.claimPromoCode(text, new Callback<PromoCode>() {
-				@Override
-				public void onSuccess(PromoCode result) {
-					dialog.dismiss();
-					Toast.makeText(getContext(), "Promo Code claimed:" + result.getCode(), Toast.LENGTH_SHORT).show();
-					showInfo(result);
-				}
-
-				@Override
-				public void onFailure(GetSocialException exception) {
-					Toast.makeText(getContext(), "Failed to claim Promo Code: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
-				}
+		public void onClick(final AlertDialog dialog, final String text) {
+			PromoCodes.claim(text, result -> {
+				dialog.dismiss();
+				Toast.makeText(getContext(), "Promo Code claimed:" + result.getCode(), Toast.LENGTH_SHORT).show();
+				showInfo(result);
+			}, exception -> {
+				Toast.makeText(getContext(), "Failed to claim Promo Code: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
 			});
 		}
 	}
