@@ -1,18 +1,18 @@
 /*
-*    	Copyright 2015-2017 GetSocial B.V.
-*
-*	Licensed under the Apache License, Version 2.0 (the "License");
-*	you may not use this file except in compliance with the License.
-*	You may obtain a copy of the License at
-*
-*    	http://www.apache.org/licenses/LICENSE-2.0
-*
-*	Unless required by applicable law or agreed to in writing, software
-*	distributed under the License is distributed on an "AS IS" BASIS,
-*	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*	See the License for the specific language governing permissions and
-*	limitations under the License.
-*/
+ *    	Copyright 2015-2017 GetSocial B.V.
+ *
+ *	Licensed under the Apache License, Version 2.0 (the "License");
+ *	you may not use this file except in compliance with the License.
+ *	You may obtain a copy of the License at
+ *
+ *    	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *	Unless required by applicable law or agreed to in writing, software
+ *	distributed under the License is distributed on an "AS IS" BASIS,
+ *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *	See the License for the specific language governing permissions and
+ *	limitations under the License.
+ */
 
 package im.getsocial.demo.adapter;
 
@@ -82,7 +82,7 @@ public class MenuAdapter extends RecyclerView.Adapter<MenuAdapter.MenuViewHolder
 			final boolean hasSubtitle = menuItem.hasSubtitle();
 			_textViewSubtitle.setText(hasSubtitle ? menuItem.getSubtitle() : null);
 			_textViewSubtitle.setVisibility(hasSubtitle ? View.VISIBLE : View.GONE);
-			
+
 			final EnabledCheck enabledCheck = menuItem.getEnabledCheck();
 			final boolean isEnabled = enabledCheck == null || enabledCheck.isOptionEnabled();
 			_view.setEnabled(isEnabled);
