@@ -1,51 +1,90 @@
 package im.getsocial.demo.adapter;
 
+import android.graphics.Color;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.RelativeLayout;
+import android.widget.Button;
 import android.widget.TextView;
 
-import java.util.Collections;
+import androidx.recyclerview.widget.RecyclerView;
+
+import java.text.DateFormat;
+import java.util.Date;
 import java.util.List;
 
-import androidx.recyclerview.widget.RecyclerView;
 import im.getsocial.demo.R;
 import im.getsocial.sdk.GetSocial;
-import im.getsocial.sdk.activities.ActivityPost;
+import im.getsocial.sdk.communities.ChatMessage;
+import im.getsocial.sdk.media.MediaAttachment;
 
-public class MessageListAdapter extends RecyclerView.Adapter<MessageListAdapter.ViewHolder> {
-	private List<ActivityPost> _items;
+public class MessageListAdapter extends RecyclerView.Adapter<MessageListAdapter.MessageViewHolder> {
+	private final List<ChatMessage> _items;
 
-	public MessageListAdapter(List<ActivityPost> items) {
+	public MessageClickListener clickListener;
+
+	public MessageListAdapter(final List<ChatMessage> items) {
 		_items = items;
-		Collections.reverse(_items);
+	}
+
+	public void appendNewMessage(ChatMessage message) {
+		_items.add(message);
+		notifyItemInserted(_items.size() - 1);
+	}
+
+	public void appendNewMessages(List<ChatMessage> newMessages) {
+		_items.addAll(newMessages);
+		notifyItemInserted(_items.size() - 1);
+	}
+
+	public void insertNewMessages(List<ChatMessage> newMessages) {
+		for(ChatMessage message : newMessages) {
+			_items.add(0, message);
+		}
+		notifyItemInserted(0);
 	}
 
 	@Override
-	public ViewHolder onCreateViewHolder(ViewGroup viewGroup, int position) {
-		View view = LayoutInflater.from(viewGroup.getContext())
-				.inflate(R.layout.list_item_message, viewGroup, false);
+	public MessageViewHolder onCreateViewHolder(final ViewGroup viewGroup, final int position) {
+		final View view = LayoutInflater.from(viewGroup.getContext())
+						.inflate(R.layout.list_item_message, viewGroup, false);
 
-		return new ViewHolder(view);
+		return new MessageViewHolder(view);
 	}
 
 	@Override
-	public void onBindViewHolder(ViewHolder viewHolder, int position) {
-		RelativeLayout.LayoutParams params = (RelativeLayout.LayoutParams)viewHolder._messageText.getLayoutParams();
+	public void onBindViewHolder(final MessageViewHolder viewHolder, final int position) {
+		viewHolder.itemView.setBackgroundColor(Color.LTGRAY);
 
-		if (GetSocial.User.getId().equals(_items.get(position).getAuthor().getId())) {
-			viewHolder._messageText.setBackgroundResource(R.drawable.message_my_background);
-			params.addRule(RelativeLayout.ALIGN_PARENT_LEFT, 0);
-			params.addRule(RelativeLayout.ALIGN_PARENT_RIGHT);
-		} else {
-			viewHolder._messageText.setBackgroundResource(R.drawable.message_user_background);
-			params.addRule(RelativeLayout.ALIGN_PARENT_RIGHT, 0);
-			params.addRule(RelativeLayout.ALIGN_PARENT_LEFT);
+		ChatMessage message = _items.get(position);
+		String senderText = "Sender: " + message.getAuthor().getDisplayName();
+		if (GetSocial.getCurrentUser().getId().equals(_items.get(position).getAuthor().getId())) {
+			senderText += " (Current)";
 		}
 
-		viewHolder._messageText.setLayoutParams(params);
-		viewHolder._messageText.setText(_items.get(position).getText());
+		viewHolder._messageSender.setText(senderText);
+		viewHolder._messageText.setText("Text:" + message.getText());
+		String date = DateFormat.getDateTimeInstance().format(new Date(message.getSentAt() * 1000));
+		viewHolder._messageDate.setText("Date:" + date);
+		String imageText = "Image: ";
+		String videoText = "Video: ";
+		if (message.getAttachments().size() > 0) {
+			MediaAttachment attachment = message.getAttachments().get(0);
+			imageText += attachment.getImageUrl();
+			videoText += attachment.getVideoUrl();
+		}
+		viewHolder._messageImageUrl.setText(imageText);
+		viewHolder._messageVideoUrl.setText(videoText);
+
+		viewHolder._showDetailsButton.setOnClickListener(new View.OnClickListener() {
+			@Override
+			public void onClick(View view) {
+				if (clickListener != null) {
+					clickListener.onShowMessageDetails(message);
+				}
+			}
+		});
+
 	}
 
 	@Override
@@ -53,12 +92,23 @@ public class MessageListAdapter extends RecyclerView.Adapter<MessageListAdapter.
 		return _items.size();
 	}
 
-	public static class ViewHolder extends RecyclerView.ViewHolder {
+	public class MessageViewHolder extends RecyclerView.ViewHolder {
+		TextView _messageSender;
 		TextView _messageText;
+		TextView _messageDate;
+		TextView _messageImageUrl;
+		TextView _messageVideoUrl;
+		Button _showDetailsButton;
 
-		ViewHolder(View view) {
+		MessageViewHolder(final View view) {
 			super(view);
-			_messageText = view.findViewById(R.id.item_message_text);
+			_messageSender = view.findViewById(R.id.message_sender);
+			_messageText = view.findViewById(R.id.message_text);
+			_messageDate = view.findViewById(R.id.message_date);
+			_messageImageUrl = view.findViewById(R.id.message_image);
+			_messageVideoUrl = view.findViewById(R.id.message_video);
+			_showDetailsButton = view.findViewById(R.id.showDetails);
 		}
+
 	}
 }
