@@ -1,9 +1,8 @@
 package im.getsocial.demo.dependencies;
 
-import im.getsocial.demo.dependencies.components.Clipboard;
 import im.getsocial.demo.dependencies.components.NotificationsManager;
-import im.getsocial.demo.utils.NotificationHandler;
 import im.getsocial.sdk.actions.ActionListener;
+import im.getsocial.sdk.notifications.OnNotificationClickedListener;
 
 public interface DependenciesContainer {
 
@@ -11,7 +10,6 @@ public interface DependenciesContainer {
 
 	ActionListener actionListener();
 
-	NotificationHandler notificationHandler();
+	OnNotificationClickedListener notificationHandler();
 
-	Clipboard clipboard();
 }
