@@ -3,14 +3,11 @@ package im.getsocial.demo.dependencies.components;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.util.Log;
-import im.getsocial.sdk.Callback;
-import im.getsocial.sdk.CompletionCallback;
 import im.getsocial.sdk.GetSocial;
-import im.getsocial.sdk.GetSocialException;
-import im.getsocial.sdk.pushnotifications.Notification;
-import im.getsocial.sdk.pushnotifications.NotificationStatus;
-import im.getsocial.sdk.pushnotifications.NotificationsCountQuery;
-import im.getsocial.sdk.pushnotifications.NotificationsQuery;
+import im.getsocial.sdk.Notifications;
+import im.getsocial.sdk.notifications.Notification;
+import im.getsocial.sdk.notifications.NotificationStatus;
+import im.getsocial.sdk.notifications.NotificationsQuery;
 
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
@@ -22,28 +19,21 @@ import java.util.Set;
 
 public final class NotificationsManager {
 
-	public interface Listener {
-		void onSync();
-	}
-
-	private boolean _isLoadingMore = false;
-
 	private final Context _context;
 	private final Set<WeakReference<Listener>> _listeners = new HashSet<>();
-
+	private final boolean _isLoadingMore = false;
 	private int _notificationsCount;
-	private List<Notification> _notifications = new ArrayList<>();
-
+	private final List<Notification> _notifications = new ArrayList<>();
 	private Set<String> _actionTypes = new HashSet<>();
 	private Set<String> _filterStatus = new HashSet<>();
 	private Set<String> _chosenTypes = new HashSet<>();
 
-	public NotificationsManager(Context context) {
+	public NotificationsManager(final Context context) {
 		_context = context;
 		loadFromPrefs();
 	}
 
-	public void saveFilter(Set<String> statuses, Set<String> chosenTypes, Set<String> actionTypes) {
+	public void saveFilter(final Set<String> statuses, final Set<String> chosenTypes, final Set<String> actionTypes) {
 		_filterStatus = new HashSet<>(statuses);
 		_chosenTypes = new HashSet<>(chosenTypes);
 		_actionTypes = actionTypes;
@@ -63,8 +53,8 @@ public final class NotificationsManager {
 		return _actionTypes;
 	}
 
-	public void addListener(Listener listener) {
-		_listeners.add(new WeakReference<Listener>(listener));
+	public void addListener(final Listener listener) {
+		_listeners.add(new WeakReference<>(listener));
 	}
 
 	public int getNewNotificationsCount() {
@@ -79,126 +69,61 @@ public final class NotificationsManager {
 		setStatus(collectAllIds(), NotificationStatus.READ);
 	}
 
-	public void setStatus(String id, String status) {
+	public void setStatus(final String id, final String status) {
 		setStatus(Collections.singletonList(id), status);
 	}
 
-	public void loadMore() {
-		if (_isLoadingMore) {
-			return;
-		}
-		_isLoadingMore = true;
-		final NotificationsQuery query = createNotificationsQuery()
-				.withFilter(NotificationsQuery.Filter.OLDER, _notifications.get(_notifications.size() - 1).getId());
-		GetSocial.User.getNotifications(query, new Callback<List<Notification>>() {
-			@Override
-			public void onSuccess(List<Notification> result) {
-				_notifications.addAll(result);
-				notifyListeners();
-				_isLoadingMore = false;
-			}
-
-			@Override
-			public void onFailure(GetSocialException exception) {
-				Log.e("Notifications", "Failed to load more: " + exception);
-				_isLoadingMore = false;
-			}
-		});
-	}
-
 	private void setStatus(final List<String> ids, final String newStatus) {
-		GetSocial.User.setNotificationsStatus(ids, newStatus, new CompletionCallback() {
-			@Override
-			public void onSuccess() {
-				updateStatuses(ids, newStatus);
-			}
-
-			@Override
-			public void onFailure(GetSocialException exception) {
-				Log.e("Notifications", "Failed to mark notification, exception: " + exception);
-			}
-		});
+		Notifications.setStatus(
+						newStatus, ids,
+						() -> updateStatuses(ids, newStatus),
+						exception -> Log.e("Notifications", "Failed to mark notification, exception: " + exception.getMessage())
+		);
 	}
 
 	public void sync() {
 		if (!GetSocial.isInitialized()) {
 			return;
 		}
-		GetSocial.User.getNotificationsCount(createNotificationsCountQuery(), new Callback<Integer>() {
-			@Override
-			public void onSuccess(Integer result) {
-				_notificationsCount = result;
-				loadNotifications();
-			}
-
-			@Override
-			public void onFailure(GetSocialException exception) {
-				Log.e("Notifications", "Failed to load notifications count: " + exception);
-				loadNotifications();
-			}
+		Notifications.getCount(createNotificationsQuery(), result -> {
+			_notificationsCount = result;
+		}, error -> {
+			Log.e("Notifications", "Failed to load notifications count: " + error);
 		});
 	}
 
-	private void loadNotifications() {
-		GetSocial.User.getNotifications(createNotificationsQuery(), new Callback<List<Notification>>() {
-			@Override
-			public void onSuccess(List<Notification> result) {
-				_notifications = new ArrayList<>(result);
-				notifyListeners();
-			}
-
-			@Override
-			public void onFailure(GetSocialException exception) {
-				Log.e("Notifications", "Failed to load notifications: " + exception);
-				notifyListeners();
-			}
-		});
-	}
-
-	private NotificationsCountQuery createNotificationsCountQuery() {
-		return NotificationsCountQuery.withStatuses(_filterStatus.toArray(new String[0]))
-				.withActions(_actionTypes.toArray(new String[0]))
-				.ofTypes(_chosenTypes.toArray(new String[0]));
-	}
-
-	private NotificationsQuery createNotificationsQuery() {
+	public NotificationsQuery createNotificationsQuery() {
 		return NotificationsQuery.withStatuses(_filterStatus.toArray(new String[0]))
-				.withActions(_actionTypes.toArray(new String[0]))
-				.ofTypes(_chosenTypes.toArray(new String[0]));
+						.withActions(_actionTypes.toArray(new String[0]))
+						.ofTypes(_chosenTypes.toArray(new String[0]));
 	}
 
-	private void updateStatuses(List<String> ids, String newStatus) {
+	private void updateStatuses(final List<String> ids, final String newStatus) {
 		for (int i = 0; i < _notifications.size(); i++) {
-			final Notification current = _notifications.get(i);
-			if (ids.contains(current.getId())) {
-				final Notification.Builder updated = Notification.builder(current.getId())
-						.withText(current.getText())
-						.withTitle(current.getTitle())
-						.withType(current.getType())
-						.withAction(current.getAction())
-						.withCreatedAt(current.getCreatedAt())
-						.withImageUrl(current.getImageUrl())
-						.withVideoUrl(current.getVideoUrl())
-						.withStatus(newStatus)
-						.addActionButtons(current.getActionButtons());
-
-				_notifications.set(i, updated.build());
+			final Notification oldNotification = _notifications.get(i);
+			if (ids.contains(oldNotification.getId())) {
+				final Notification notification = new Notification(
+								oldNotification.getId(), oldNotification.getActionButtons(), newStatus, oldNotification.getType(), oldNotification.getCreatedAt(),
+								oldNotification.getTitle(), oldNotification.getText(), oldNotification.getAction(),
+								oldNotification.getAttachment(), oldNotification.getSender(), oldNotification.getCustomization()
+				);
+				_notifications.set(i, notification);
 			}
 		}
 		notifyListeners();
 	}
 
 	private List<String> collectAllIds() {
-		List<String> ids = new ArrayList<>(_notifications.size());
-		for (Notification notification : _notifications) {
+		final List<String> ids = new ArrayList<>(_notifications.size());
+		for (final Notification notification : _notifications) {
 			ids.add(notification.getId());
 		}
 		return ids;
 	}
 
 	private void notifyListeners() {
-		for (Iterator<WeakReference<Listener>> iterator = _listeners.iterator(); iterator.hasNext();) {
-			Listener listener = iterator.next().get();
+		for (final Iterator<WeakReference<Listener>> iterator = _listeners.iterator(); iterator.hasNext(); ) {
+			final Listener listener = iterator.next().get();
 			if (listener != null) {
 				listener.onSync();
 			} else {
@@ -209,18 +134,22 @@ public final class NotificationsManager {
 
 	private void saveToPrefs() {
 		_context.getSharedPreferences("notifications", Context.MODE_PRIVATE)
-				.edit()
-				.putStringSet("status", _filterStatus)
-				.putStringSet("types", _chosenTypes)
-				.putStringSet("actions", _actionTypes)
-				.apply();
+						.edit()
+						.putStringSet("status", _filterStatus)
+						.putStringSet("types", _chosenTypes)
+						.putStringSet("actions", _actionTypes)
+						.apply();
 	}
 
 	private void loadFromPrefs() {
 		final SharedPreferences sharedPreferences = _context.getSharedPreferences("notifications", Context.MODE_PRIVATE);
 
-		_chosenTypes = sharedPreferences.getStringSet("types", Collections.<String>emptySet());
-		_filterStatus = sharedPreferences.getStringSet("status", Collections.<String>emptySet());
-		_actionTypes = sharedPreferences.getStringSet("actions", Collections.<String>emptySet());
+		_chosenTypes = sharedPreferences.getStringSet("types", Collections.emptySet());
+		_filterStatus = sharedPreferences.getStringSet("status", Collections.emptySet());
+		_actionTypes = sharedPreferences.getStringSet("actions", Collections.emptySet());
+	}
+
+	public interface Listener {
+		void onSync();
 	}
 }
