@@ -1,86 +1,139 @@
 /*
-*    	Copyright 2015-2017 GetSocial B.V.
-*
-*	Licensed under the Apache License, Version 2.0 (the "License");
-*	you may not use this file except in compliance with the License.
-*	You may obtain a copy of the License at
-*
-*    	http://www.apache.org/licenses/LICENSE-2.0
-*
-*	Unless required by applicable law or agreed to in writing, software
-*	distributed under the License is distributed on an "AS IS" BASIS,
-*	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*	See the License for the specific language governing permissions and
-*	limitations under the License.
-*/
+ *    	Copyright 2015-2017 GetSocial B.V.
+ *
+ *	Licensed under the Apache License, Version 2.0 (the "License");
+ *	you may not use this file except in compliance with the License.
+ *	You may obtain a copy of the License at
+ *
+ *    	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *	Unless required by applicable law or agreed to in writing, software
+ *	distributed under the License is distributed on an "AS IS" BASIS,
+ *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *	See the License for the specific language governing permissions and
+ *	limitations under the License.
+ */
 
 package im.getsocial.demo.fragment;
 
 import android.app.AlertDialog;
-import android.content.DialogInterface;
-import android.text.TextUtils;
 import android.widget.Toast;
 
 import im.getsocial.demo.Utils;
 import im.getsocial.demo.adapter.MenuItem;
-import im.getsocial.demo.dialog.action_dialog.ActionDialog;
-import im.getsocial.sdk.Callback;
+import im.getsocial.sdk.Communities;
 import im.getsocial.sdk.ErrorCode;
 import im.getsocial.sdk.GetSocial;
-import im.getsocial.sdk.GetSocialException;
 import im.getsocial.sdk.actions.Action;
 import im.getsocial.sdk.actions.ActionListener;
-import im.getsocial.sdk.actions.ActionTypes;
-import im.getsocial.sdk.activities.ActivitiesQuery;
-import im.getsocial.sdk.activities.ActivityPost;
-import im.getsocial.sdk.pushnotifications.Notification;
-import im.getsocial.sdk.ui.AvatarClickListener;
+import im.getsocial.sdk.common.PagingQuery;
+import im.getsocial.sdk.communities.ActivitiesQuery;
+import im.getsocial.sdk.communities.GetSocialActivity;
+import im.getsocial.sdk.communities.PollStatus;
+import im.getsocial.sdk.communities.UserId;
+import im.getsocial.sdk.media.MediaAttachment;
 import im.getsocial.sdk.ui.CustomErrorMessageProvider;
-import im.getsocial.sdk.ui.GetSocialUi;
-import im.getsocial.sdk.ui.MentionClickListener;
-import im.getsocial.sdk.ui.TagClickListener;
 import im.getsocial.sdk.ui.UiAction;
-import im.getsocial.sdk.ui.UiActionListener;
 import im.getsocial.sdk.ui.ViewStateListener;
-import im.getsocial.sdk.ui.activities.ActionButtonListener;
-import im.getsocial.sdk.ui.activities.ActivityFeedViewBuilder;
-import im.getsocial.sdk.usermanagement.PublicUser;
+import im.getsocial.sdk.ui.communities.ActivityDetailsViewBuilder;
+import im.getsocial.sdk.ui.communities.ActivityFeedViewBuilder;
 
 import java.util.Arrays;
 import java.util.List;
 
-public class ActivitiesFragment extends BaseListFragment implements ActionListener, ActionButtonListener {
+public class ActivitiesFragment extends BaseListFragment implements ActionListener {
 
-	public static final String CUSTOM_FEED_NAME = "DemoFeed";
 	private static final List<UiAction> FORBIDDEN_FOR_ANONYMOUS = Arrays.asList(UiAction.LIKE_ACTIVITY, UiAction.LIKE_COMMENT, UiAction.POST_ACTIVITY, UiAction.POST_COMMENT);
 
 	@Override
 	protected List<MenuItem> createListData() {
 		return Arrays.asList(
-				new MenuItem.Builder("Global Activity Feed")
-						.withAction(new OpenDefaultGlobalFeedAction())
+						new MenuItem.Builder("Timeline with Handlers")
+										.withAction(new OpenTimelineWithHandlers())
+										.build(),
+						new MenuItem.Builder("My Activity Feed")
+										.withAction(new OpenMyFeed())
+										.withSubtitle("Is not available in phase 1")
+										.withEnabledCheck(() -> false)
+										.build(),
+						new MenuItem.Builder("Demo Feed")
+										.withAction(new OpenTopicFeedAction("DemoFeed"))
+										.build(),
+						new MenuItem.Builder("Timeline")
+										.withAction(new OpenTimeline())
+										.build(),
+						new MenuItem.Builder("Open Activity Details From Timeline")
+										.withAction(new OpenActivityDetailsAction(true))
+										.build(),
+						new MenuItem.Builder("Open Activity Details From Timeline(without feed view)")
+										.withAction(new OpenActivityDetailsAction(false))
+										.build(),
+						MenuItem.builder("Update Last Activity")
+										.withAction(new UpdateActivityAction())
+										.build(),
+						MenuItem.builder("All my posts everywhere")
+										.withAction(new MenuItem.Action() {
+											@Override
+											public void execute() {
+												ActivityFeedViewBuilder.create(ActivitiesQuery.everywhere().byUser(UserId.currentUser()))
+																.show();
+											}
+										})
+										.build(),
+						MenuItem.builder("All my posts in all topics")
+								.withAction(new MenuItem.Action() {
+									@Override
+									public void execute() {
+										ActivityFeedViewBuilder.create(ActivitiesQuery.inAllTopics().byUser(UserId.currentUser()))
+												.show();
+									}
+								})
+								.build(),
+						MenuItem.builder("All my posts in all groups")
+								.withAction(new MenuItem.Action() {
+									@Override
+									public void execute() {
+										ActivityFeedViewBuilder.create(ActivitiesQuery.inAllGroups().byUser(UserId.currentUser()))
+												.show();
+									}
+								})
+								.build(),
+						MenuItem.builder("My Feed")
+										.withAction(new MenuItem.Action() {
+											@Override
+											public void execute() {
+												ActivityFeedViewBuilder.create(ActivitiesQuery.feedOf(UserId.currentUser()))
+																.show();
+											}
+										})
+										.build(),
+						navigationListItem("Post to timeline", PostActivityFragment.class),
+						MenuItem.builder("Create Poll in User's feed")
+								.withAction(new MenuItem.Action() {
+									@Override
+									public void execute() {
+										addContentFragment(CreatePollFragment.postToTimeline());
+									}
+								})
+								.build(),
+						navigationListItem("Reactions", FeedFragment.class),
+						MenuItem.builder("Topic Activity Details (569141942483009835)")
+							.withAction(new MenuItem.Action() {
+								@Override
+								public void execute() {
+									ActivityDetailsViewBuilder.create("569141942483009835")
+										.show();
+							}
+						})
 						.build(),
-				new MenuItem.Builder("Global Activity Feed With Custom Handlers")
-						.withAction(new OpenGlobalFeedAction())
-						.build(),
-				new MenuItem.Builder(String.format("Custom Activity Feed (%s)", CUSTOM_FEED_NAME))
-						.withAction(new OpenCustomFeedAction(CUSTOM_FEED_NAME))
-						.build(),
-				new MenuItem.Builder("My Global Activity Feed")
-						.withAction(new OpenMyGlobalFeedAction())
-						.build(),
-				new MenuItem.Builder("My Friends Global Feed")
-						.withAction(new OpenMyFriendsGlobalFeedAction())
-						.build(),
-				new MenuItem.Builder("My Custom Activity Feed")
-						.withAction(new OpenMyCustomFeedAction(CUSTOM_FEED_NAME))
-						.build(),
-				navigationListItem("Post Activity", PostActivityFragment.class),
-				new MenuItem.Builder("Open Activity Details From Global Feed")
-						.withAction(new OpenActivityDetailsAction(true))
-						.build(),
-				new MenuItem.Builder("Open Activity Details From Global Feed(without feed view)")
-						.withAction(new OpenActivityDetailsAction(false))
+						MenuItem.builder("Group Activity Details (504205474833540031)")
+								.withAction(new MenuItem.Action() {
+								@Override
+								public void execute() {
+									ActivityDetailsViewBuilder.create("504205474833540031")
+											.show();
+								}
+						})
 						.build()
 		);
 	}
@@ -95,173 +148,70 @@ public class ActivitiesFragment extends BaseListFragment implements ActionListen
 		return "Activities";
 	}
 
-	private void openGlobalFeedForTag(String tag) {
-		GetSocialUi.createGlobalActivityFeedView()
-				.setWindowTitle(String.format("Search #%s", tag))
-				.setFilterByTags(tag)
-				.setReadOnly(true)
-				.show();
+	private void openGlobalFeedForTag(final String tag) {
+		ActivityFeedViewBuilder.create(ActivitiesQuery.everywhere().withTag(tag))
+						.setWindowTitle(String.format("Search #%s", tag))
+						.show();
 	}
 
-	private void getUserAndShowActionDialog(String mention) {
-		if (mention.equals(MentionClickListener.APP_SHORTCUT)) {
-			Toast.makeText(getContext(), "Application mention clicked.", Toast.LENGTH_SHORT).show();
-			return;
-		}
-		GetSocial.getUserById(mention, new Callback<PublicUser>() {
-			@Override
-			public void onSuccess(PublicUser user) {
-				showUserActionDialog(user);
-			}
 
-			@Override
-			public void onFailure(GetSocialException exception) {
-				_log.logErrorAndToast("Failed to get user: " + exception.getMessage());
-			}
-		});
-	}
-
-	private void showUserActionDialog(final PublicUser user) {
-		ActionDialog actionDialog = new ActionDialog(getContext());
-
-		if (isCurrentUser(user)) {
-			actionDialog.addAction(new ActionDialog.Action("Show User Feed") {
+	private void showCurrentUserFeed() {
+		ActivityFeedViewBuilder builder = ActivityFeedViewBuilder.create(ActivitiesQuery.feedOf(UserId.currentUser()))
+						.setWindowTitle("My Feed")
+						.setActionListener(actionListener());
+		if (Utils.isCustomErrorMessageEnabled(getContext())) {
+			builder.setCustomErrorMessageProvider(new CustomErrorMessageProvider() {
 				@Override
-				public void execute() {
-					showGlobalFeedForCurrentUser();
+				public String onError(int errorCode, String errorMessage) {
+					if (errorCode == ErrorCode.ACTIVITY_REJECTED) {
+						return "Be careful what you say :)";
+					}
+					return errorMessage;
 				}
 			});
-			actionDialog.show();
-		} else {
-			actionDialog.setTitle("User " + user.getDisplayName());
-			actionDialog.addAction(new ActionDialog.Action("Show User Feed") {
-				@Override
-				public void execute() {
-					showGlobalFeedForOtherUser(user);
-				}
-			});
-			checkIfFriendsAndShowDialog(user, actionDialog);
 		}
-	}
-
-	private void showGlobalFeedForCurrentUser() {
-		showGlobalFeedForUser(GetSocial.User.getId(), "My Global Feed");
-	}
-
-	private void showGlobalFeedForOtherUser(PublicUser user) {
-		showGlobalFeedForUser(user.getId(), user.getDisplayName() + " Global Feed");
-	}
-
-	private void showGlobalFeedForUser(String id, String title) {
-		GetSocialUi.createGlobalActivityFeedView()
-				.setFilterByUser(id)
-				.setWindowTitle(title)
-				.setReadOnly(true)
-				.setButtonActionListener(actionListener())
-				.setActionListener(actionListener())
-				.show();
-	}
-
-	private boolean isCurrentUser(PublicUser user) {
-		return user.getId().equals(GetSocial.User.getId());
-	}
-
-	private void checkIfFriendsAndShowDialog(final PublicUser user, final ActionDialog actionDialog) {
-		GetSocial.User.isFriend(user.getId(), new Callback<Boolean>() {
-			@Override
-			public void onSuccess(Boolean isFriend) {
-				if (isFriend) {
-					actionDialog.addAction(new ActionDialog.Action("Remove from Friends") {
-						@Override
-						public void execute() {
-							removeFriend(user);
-						}
-					});
-				} else {
-					actionDialog.addAction(new ActionDialog.Action("Add to Friends") {
-						@Override
-						public void execute() {
-							addFriend(user);
-						}
-					});
-				}
-				actionDialog.show();
-			}
-
-			@Override
-			public void onFailure(GetSocialException exception) {
-				_log.logErrorAndToast("Failed to check if friend: " + exception.getMessage());
-			}
-		});
-	}
-
-	private void addFriend(final PublicUser user) {
-		GetSocial.User.addFriend(user.getId(), new Callback<Integer>() {
-			@Override
-			public void onSuccess(Integer integer) {
-				Toast.makeText(getContext(), user.getDisplayName() + " is now your friend!", Toast.LENGTH_SHORT).show();
-			}
-
-			@Override
-			public void onFailure(GetSocialException exception) {
-				_log.logErrorAndToast("Failed to add friend: " + exception.getMessage());
-			}
-		});
-	}
-
-	private void removeFriend(final PublicUser user) {
-		GetSocial.User.removeFriend(user.getId(), new Callback<Integer>() {
-			@Override
-			public void onSuccess(Integer integer) {
-				Toast.makeText(getContext(), user.getDisplayName() + " is not your friend anymore!", Toast.LENGTH_SHORT).show();
-			}
-
-			@Override
-			public void onFailure(GetSocialException exception) {
-				_log.logErrorAndToast("Failed to remove friend: " + exception.getMessage());
-			}
-		});
+		builder.show();
 	}
 
 	@Override
-	public boolean handleAction(Action action) {
-		return _activityListener.dependencies().actionListener().handleAction(action);
+	public void handleAction(final Action action) {
+		_activityListener.dependencies().actionListener().handleAction(action);
 	}
 
-	@Override
-	public void onButtonClicked(String action, ActivityPost post) {
-		if ("CloseAndRestore".equalsIgnoreCase(action)) {
-			GetSocialUi.closeView(true);
-			new AlertDialog.Builder(getContext())
-					.setPositiveButton("Restore View", new DialogInterface.OnClickListener() {
-						@Override
-						public void onClick(DialogInterface dialogInterface, int i) {
-							GetSocialUi.restoreView();
-						}
-					})
-					.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-						@Override
-						public void onClick(DialogInterface dialog, int which) {
-
-						}
-					}).show();
-		}
-
-		if (!TextUtils.isEmpty(action)) {
-			Toast.makeText(getContext(), "Activity Feed button clicked, action: " + action, Toast.LENGTH_SHORT).show();
-		}
-	}
-
-	private <T extends ActionButtonListener & ActionListener> T actionListener() {
+	private <T extends ActionListener> T actionListener() {
 		return (T) this;
 	}
 
-	private class OpenDefaultGlobalFeedAction implements MenuItem.Action {
+	private class OpenTimelineWithHandlers implements MenuItem.Action {
 
 		@Override
 		public void execute() {
-			ActivityFeedViewBuilder builder = GetSocialUi.createGlobalActivityFeedView();
-			if (Utils.isCustomErrorMesageEnabled(getContext())) {
+			ActivityFeedViewBuilder builder = ActivityFeedViewBuilder.create(ActivitiesQuery.timeline())
+							.setActionListener(actionListener())
+							.setViewStateListener(new ViewStateListener() {
+								@Override
+								public void onOpen() {
+									_log.logInfoAndToast("Global feed was opened");
+								}
+
+								@Override
+								public void onClose() {
+									_log.logInfoAndToast("Global feed was closed");
+								}
+							})
+							.setUiActionListener((action, pendingAction) -> {
+								final String actionDescription = action.name().replace("_", " ").toLowerCase();
+								_log.logInfoAndToast("User is going to " + actionDescription);
+								if (GetSocial.getCurrentUser().isAnonymous() && FORBIDDEN_FOR_ANONYMOUS.contains(action)) {
+									showAuthorizeUserDialogForPendingAction(actionDescription, pendingAction);
+								} else {
+									pendingAction.proceed();
+								}
+							})
+							.setAvatarClickListener(ActivitiesFragment.this::showUserActionDialog)
+							.setMentionClickListener(mention -> getUserAndShowActionDialog(mention))
+							.setTagClickListener(ActivitiesFragment.this::openGlobalFeedForTag);
+			if (Utils.isCustomErrorMessageEnabled(getContext())) {
 				builder.setCustomErrorMessageProvider(new CustomErrorMessageProvider() {
 					@Override
 					public String onError(int errorCode, String errorMessage) {
@@ -276,55 +226,50 @@ public class ActivitiesFragment extends BaseListFragment implements ActionListen
 		}
 	}
 
-	private class OpenGlobalFeedAction implements MenuItem.Action {
+	private class OpenTopicFeedAction implements MenuItem.Action {
+
+		private final String _feed;
+
+		OpenTopicFeedAction(final String feed) {
+			_feed = feed;
+		}
 
 		@Override
 		public void execute() {
-			ActivityFeedViewBuilder builder = GetSocialUi.createGlobalActivityFeedView()
-					.setButtonActionListener(actionListener())
-					.setActionListener(actionListener())
-					.setViewStateListener(new ViewStateListener() {
-						@Override
-						public void onOpen() {
-							_log.logInfoAndToast("Global feed was opened");
-						}
+			ActivityFeedViewBuilder builder =  ActivityFeedViewBuilder.create(ActivitiesQuery.activitiesInTopic(_feed))
+							.setActionListener(actionListener())
+							.setAvatarClickListener(ActivitiesFragment.this::showUserActionDialog);
 
-						@Override
-						public void onClose() {
-							_log.logInfoAndToast("Global feed was closed");
-						}
-					})
-					.setUiActionListener(new UiActionListener() {
-						@Override
-						public void onUiAction(UiAction action, UiAction.Pending pendingAction) {
-							final String actionDescription = action.name().replace("_", " ").toLowerCase();
-							_log.logInfoAndToast("User is going to " + actionDescription);
-							if (GetSocial.User.isAnonymous() && FORBIDDEN_FOR_ANONYMOUS.contains(action)) {
-								showAuthorizeUserDialogForPendingAction(actionDescription, pendingAction);
-							} else {
-								pendingAction.proceed();
+							if (Utils.isCustomErrorMessageEnabled(getContext())) {
+								builder.setCustomErrorMessageProvider(new CustomErrorMessageProvider() {
+									@Override
+									public String onError(int errorCode, String errorMessage) {
+										if (errorCode == ErrorCode.ACTIVITY_REJECTED) {
+											return "Be careful what you say :)";
+										}
+										return errorMessage;
+									}
+								});
 							}
-						}
-					})
-					.setAvatarClickListener(new AvatarClickListener() {
-						@Override
-						public void onAvatarClicked(PublicUser user) {
-							showUserActionDialog(user);
-						}
-					})
-					.setMentionClickListener(new MentionClickListener() {
-						@Override
-						public void onMentionClicked(String mention) {
-							getUserAndShowActionDialog(mention);
-						}
-					})
-					.setTagClickListener(new TagClickListener() {
-						@Override
-						public void onTagClicked(String tag) {
-							openGlobalFeedForTag(tag);
-						}
-					});
-			if (Utils.isCustomErrorMesageEnabled(getContext())) {
+							builder.show();
+		}
+	}
+
+	private class OpenMyFeed implements MenuItem.Action {
+
+		@Override
+		public void execute() {
+			showCurrentUserFeed();
+		}
+	}
+
+	private class OpenTimeline implements MenuItem.Action {
+
+		@Override
+		public void execute() {
+			ActivityFeedViewBuilder builder = ActivityFeedViewBuilder.create(ActivitiesQuery.timeline())
+							.setActionListener(actionListener());
+			if (Utils.isCustomErrorMessageEnabled(getContext())) {
 				builder.setCustomErrorMessageProvider(new CustomErrorMessageProvider() {
 					@Override
 					public String onError(int errorCode, String errorMessage) {
@@ -335,85 +280,6 @@ public class ActivitiesFragment extends BaseListFragment implements ActionListen
 					}
 				});
 			}
-			builder.show();
-		}
-	}
-
-	private class OpenCustomFeedAction implements MenuItem.Action {
-
-		private final String _feed;
-
-		OpenCustomFeedAction(String feed) {
-			_feed = feed;
-		}
-
-		@Override
-		public void execute() {
-			ActivityFeedViewBuilder builder = GetSocialUi.createActivityFeedView(_feed)
-					.setButtonActionListener(actionListener())
-					.setActionListener(actionListener());
-			if (Utils.isCustomErrorMesageEnabled(getContext())) {
-				builder.setCustomErrorMessageProvider(new CustomErrorMessageProvider() {
-					@Override
-					public String onError(int errorCode, String errorMessage) {
-						if (errorCode == ErrorCode.ACTIVITY_REJECTED) {
-							return "Be careful what you say :)";
-						}
-						return errorMessage;
-					}
-				});
-			}
-			builder.show();
-		}
-	}
-
-	private class OpenMyGlobalFeedAction implements MenuItem.Action {
-
-		@Override
-		public void execute() {
-			showGlobalFeedForCurrentUser();
-		}
-	}
-
-	private class OpenMyFriendsGlobalFeedAction implements MenuItem.Action {
-
-		@Override
-		public void execute() {
-			ActivityFeedViewBuilder builder = GetSocialUi.createGlobalActivityFeedView()
-					.setButtonActionListener(actionListener())
-					.setActionListener(actionListener())
-					.setShowFriendsFeed(true);
-			if (Utils.isCustomErrorMesageEnabled(getContext())) {
-				builder.setCustomErrorMessageProvider(new CustomErrorMessageProvider() {
-					@Override
-					public String onError(int errorCode, String errorMessage) {
-						if (errorCode == ErrorCode.ACTIVITY_REJECTED) {
-							return "Be careful what you say :)";
-						}
-						return errorMessage;
-					}
-				});
-			}
-			builder.show();
-		}
-	}
-
-	private class OpenMyCustomFeedAction implements MenuItem.Action {
-
-		private final String _feed;
-
-		OpenMyCustomFeedAction(String feed) {
-			_feed = feed;
-		}
-
-		@Override
-		public void execute() {
-			ActivityFeedViewBuilder builder = GetSocialUi.createActivityFeedView(_feed)
-					.setFilterByUser(GetSocial.User.getId())
-					.setWindowTitle("My Custom Feed")
-					.setReadOnly(false)
-					.setButtonActionListener(actionListener())
-					.setActionListener(actionListener());
 			builder.show();
 		}
 	}
@@ -422,71 +288,97 @@ public class ActivitiesFragment extends BaseListFragment implements ActionListen
 
 		private final boolean _showFeed;
 
-		OpenActivityDetailsAction(boolean showFeed) {
+		OpenActivityDetailsAction(final boolean showFeed) {
 			_showFeed = showFeed;
 		}
 
 		@Override
 		public void execute() {
-			GetSocial.getActivities(ActivitiesQuery.postsForGlobalFeed().withLimit(5), new Callback<List<ActivityPost>>() {
-				@Override
-				public void onSuccess(final List<ActivityPost> activityPosts) {
-					if (activityPosts.isEmpty()) {
-						Toast.makeText(getContext(), "No activities in global feed", Toast.LENGTH_SHORT).show();
-						return;
-					}
-					final String[] activityContents = new String[activityPosts.size()];
-					for (int i = 0; i < activityContents.length; i++) {
-						final ActivityPost activityPost = activityPosts.get(i);
-						if (activityPost.hasText()) {
-							activityContents[i] = activityPost.getText();
-						} else if (activityPost.hasImage()) {
-							activityContents[i] = "[IMAGE] " + activityPost.getImageUrl();
-						} else if (activityPost.hasButton()) {
-							activityContents[i] = "[BUTTON] " + activityPost.getButtonTitle();
+			Communities.getActivities(new PagingQuery<>(ActivitiesQuery.timeline()).withLimit(5), result -> {
+				final List<GetSocialActivity> getSocialActivities = result.getEntries();
+				if (getSocialActivities.isEmpty()) {
+					Toast.makeText(getContext(), "No activities in timeline", Toast.LENGTH_SHORT).show();
+					return;
+				}
+				final String[] activityContents = new String[getSocialActivities.size()];
+				for (int i = 0; i < activityContents.length; i++) {
+					final GetSocialActivity activity = getSocialActivities.get(i);
+					if (activity.getText() != null) {
+						activityContents[i] = activity.getText();
+					} else if (activity.getAttachments().size() > 0) {
+						final MediaAttachment attachment = activity.getAttachments().get(0);
+						if (attachment.getImageUrl() != null) {
+							activityContents[i] = "[IMAGE] " + attachment.getImageUrl();
 						} else {
-							activityContents[i] = activityPost.toString();
+							activityContents[i] = "[VIDEO] " + attachment.getVideoUrl();
 						}
+					} else {
+						activityContents[i] = activity.toString();
 					}
-					new AlertDialog.Builder(getContext())
-							.setItems(activityContents, new DialogInterface.OnClickListener() {
-								@Override
-								public void onClick(DialogInterface dialog, int which) {
+				}
+				new AlertDialog.Builder(getContext())
+								.setItems(activityContents, (dialog, which) -> {
 									dialog.dismiss();
-									String activityId = activityPosts.get(which).getId();
-									GetSocialUi.createActivityDetailsView(activityId)
-											.setViewStateListener(new ViewStateListener() {
-												@Override
-												public void onOpen() {
-													Toast.makeText(getContext(), "Activity details view opened", Toast.LENGTH_SHORT).show();
-												}
+									String activityId = getSocialActivities.get(which).getId();
+									ActivityDetailsViewBuilder.create(activityId)
+													.setViewStateListener(new ViewStateListener() {
+														@Override
+														public void onOpen() {
+															Toast.makeText(getContext(), "Activity details view opened", Toast.LENGTH_SHORT).show();
+														}
 
-												@Override
-												public void onClose() {
-													Toast.makeText(getContext(), "Activity details view closed", Toast.LENGTH_SHORT).show();
-												}
-											})
-											.setButtonActionListener(actionListener())
-											.setActionListener(actionListener())
-											.setWindowTitle("Activity Details")
-											.setUiActionListener(new UiActionListener() {
-												@Override
-												public void onUiAction(UiAction action, UiAction.Pending pendingAction) {
-													Toast.makeText(getContext(), "Action done: " + action.name(), Toast.LENGTH_SHORT).show();
-													pendingAction.proceed();
-												}
-											})
-											.setShowActivityFeedView(_showFeed)
-											.show();
-								}
-							}).show();
-				}
+														@Override
+														public void onClose() {
+															Toast.makeText(getContext(), "Activity details view closed", Toast.LENGTH_SHORT).show();
+														}
+													})
+													.setActionListener(actionListener())
+													.setWindowTitle("Activity Details")
+													.setUiActionListener((action, pendingAction) -> {
+														Toast.makeText(getContext(), "Action done: " + action.name(), Toast.LENGTH_SHORT).show();
+														pendingAction.proceed();
+													})
+													.setShowActivityFeedView(_showFeed)
+													.show();
+								}).show();
+			}, exception -> _log.logErrorAndToast("Failed to load activities, error: " + exception.getMessage()));
+		}
+	}
 
-				@Override
-				public void onFailure(GetSocialException exception) {
-					_log.logErrorAndToast("Failed to load activities, error: " + exception.getMessage());
+	private class UpdateActivityAction implements MenuItem.Action {
+
+		@Override
+		public void execute() {
+			Communities.getActivities(new PagingQuery<>(ActivitiesQuery.everywhere().byUser(UserId.currentUser()).withPollStatus(PollStatus.WITHOUT_POLL)).withLimit(5), result -> {
+				final List<GetSocialActivity> getSocialActivities = result.getEntries();
+				if (getSocialActivities.isEmpty()) {
+					Toast.makeText(getContext(), "You haven't recently posted anywhere.", Toast.LENGTH_SHORT).show();
+					return;
 				}
-			});
+				final String[] activityContents = new String[getSocialActivities.size()];
+				for (int i = 0; i < activityContents.length; i++) {
+					final GetSocialActivity activity = getSocialActivities.get(i);
+					if (activity.getText() != null) {
+						activityContents[i] = activity.getText();
+					} else if (activity.getAttachments().size() > 0) {
+						final MediaAttachment attachment = activity.getAttachments().get(0);
+						if (attachment.getImageUrl() != null) {
+							activityContents[i] = "[IMAGE] " + attachment.getImageUrl();
+						} else {
+							activityContents[i] = "[VIDEO] " + attachment.getVideoUrl();
+						}
+					} else {
+						activityContents[i] = activity.toString();
+					}
+				}
+				new AlertDialog.Builder(getContext())
+								.setItems(activityContents, (dialog, which) -> {
+									dialog.dismiss();
+									String activityId = getSocialActivities.get(which).getId();
+
+									addContentFragment(PostActivityFragment.updateActivity(activityId));
+								}).show();
+			}, exception -> _log.logErrorAndToast("Failed to load activities, error: " + exception.getMessage()));
 		}
 	}
 }
