@@ -1,18 +1,18 @@
 /*
-*    	Copyright 2015-2017 GetSocial B.V.
-*
-*	Licensed under the Apache License, Version 2.0 (the "License");
-*	you may not use this file except in compliance with the License.
-*	You may obtain a copy of the License at
-*
-*    	http://www.apache.org/licenses/LICENSE-2.0
-*
-*	Unless required by applicable law or agreed to in writing, software
-*	distributed under the License is distributed on an "AS IS" BASIS,
-*	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*	See the License for the specific language governing permissions and
-*	limitations under the License.
-*/
+ *    	Copyright 2015-2017 GetSocial B.V.
+ *
+ *	Licensed under the Apache License, Version 2.0 (the "License");
+ *	you may not use this file except in compliance with the License.
+ *	You may obtain a copy of the License at
+ *
+ *    	http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *	Unless required by applicable law or agreed to in writing, software
+ *	distributed under the License is distributed on an "AS IS" BASIS,
+ *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *	See the License for the specific language governing permissions and
+ *	limitations under the License.
+ */
 
 package im.getsocial.demo.fragment;
 
@@ -40,12 +40,12 @@ public abstract class BaseListFragment extends BaseFragment {
 	private ViewContainer _viewContainer;
 
 	@Override
-	public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+	public View onCreateView(final LayoutInflater inflater, @Nullable final ViewGroup container, @Nullable final Bundle savedInstanceState) {
 		return inflater.inflate(R.layout.fragment_menu, container, false);
 	}
 
 	@Override
-	public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+	public void onViewCreated(final View view, @Nullable final Bundle savedInstanceState) {
 		super.onViewCreated(view, savedInstanceState);
 		_viewContainer = new ViewContainer(view);
 	}
@@ -56,12 +56,12 @@ public abstract class BaseListFragment extends BaseFragment {
 		invalidateList();
 	}
 
-	protected MenuItem navigationListItem(String name, Class<? extends Fragment> fragmentClass) {
+	protected MenuItem navigationListItem(final String name, final Class<? extends Fragment> fragmentClass) {
 		return navigationListItem(name, fragmentClass, null);
 	}
 
-	protected MenuItem navigationListItem(String name, Class<? extends Fragment> fragmentClass, @Nullable NavigationItemDecorator decorator) {
-		MenuItem.Builder builder = MenuItem.builder(name).withAction(new OpenFragmentAction(fragmentClass));
+	protected MenuItem navigationListItem(final String name, final Class<? extends Fragment> fragmentClass, @Nullable final NavigationItemDecorator decorator) {
+		final MenuItem.Builder builder = MenuItem.builder(name).withAction(new OpenFragmentAction(fragmentClass));
 		if (decorator != null) {
 			decorator.decorate(builder);
 		}
@@ -78,22 +78,26 @@ public abstract class BaseListFragment extends BaseFragment {
 
 	protected abstract List<MenuItem> createListData();
 
+	protected interface NavigationItemDecorator {
+		void decorate(MenuItem.Builder builder);
+	}
+
 	class ViewContainer {
-		private final LinearLayoutManager _linearLayoutManager;
+
 		@BindView(recyclerView)
 		RecyclerView _recyclerView;
 
-		ViewContainer(View view) {
+		ViewContainer(final View view) {
 			ButterKnife.bind(this, view);
 
-			_linearLayoutManager = new LinearLayoutManager(getContext());
-			_recyclerView.setLayoutManager(_linearLayoutManager);
+			final LinearLayoutManager linearLayoutManager = new LinearLayoutManager(getContext());
+			_recyclerView.setLayoutManager(linearLayoutManager);
 
-			DividerItemDecoration dividerItemDecoration = new DividerItemDecoration(getContext(),
-					_linearLayoutManager.getOrientation());
+			final DividerItemDecoration dividerItemDecoration = new DividerItemDecoration(getContext(),
+							linearLayoutManager.getOrientation());
 			_recyclerView.addItemDecoration(dividerItemDecoration);
 
-			MenuAdapter adapter = new MenuAdapter(createListData());
+			final MenuAdapter adapter = new MenuAdapter(createListData());
 			_recyclerView.setAdapter(adapter);
 		}
 	}
@@ -102,7 +106,7 @@ public abstract class BaseListFragment extends BaseFragment {
 
 		private final Class<? extends Fragment> _fragmentClass;
 
-		OpenFragmentAction(Class<? extends Fragment> fragmentClass) {
+		OpenFragmentAction(final Class<? extends Fragment> fragmentClass) {
 			_fragmentClass = fragmentClass;
 		}
 
@@ -110,13 +114,9 @@ public abstract class BaseListFragment extends BaseFragment {
 		public void execute() {
 			try {
 				addContentFragment(_fragmentClass.newInstance());
-			} catch (Exception exception) {
+			} catch (final Exception exception) {
 				throw new RuntimeException(exception);
 			}
 		}
-	}
-
-	protected interface NavigationItemDecorator {
-		void decorate(MenuItem.Builder builder);
 	}
 }
