@@ -1,59 +1,50 @@
 package im.getsocial.demo.fragment;
 
 import android.app.AlertDialog;
-import android.app.DatePickerDialog;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.DatePicker;
 import android.widget.EditText;
 import android.widget.LinearLayout;
-import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.TimePicker;
 import android.widget.Toast;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 import im.getsocial.demo.R;
 import im.getsocial.demo.utils.DynamicUi;
-import im.getsocial.sdk.Callback;
-import im.getsocial.sdk.GetSocial;
-import im.getsocial.sdk.GetSocialException;
-import im.getsocial.sdk.promocodes.PromoCode;
-import im.getsocial.sdk.promocodes.PromoCodeBuilder;
+import im.getsocial.sdk.PromoCodes;
+import im.getsocial.sdk.promocodes.PromoCodeContent;
 
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
-import java.util.TimeZone;
 
 public class CreatePromoCodeFragment extends BaseFragment {
 
+	private final List<DynamicUi.DynamicInputHolder> _customData = new ArrayList<>();
 	private ViewContainer _viewContainer;
-
 	@Nullable
 	private Date _startTime;
 	@Nullable
 	private Date _endTime;
-
 	private boolean _creating = false;
-
-	private final List<DynamicUi.DynamicInputHolder> _customData = new ArrayList<>();
 
 	@Nullable
 	@Override
-	public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+	public View onCreateView(@NonNull final LayoutInflater inflater, @Nullable final ViewGroup container, @Nullable final Bundle savedInstanceState) {
 		return inflater.inflate(R.layout.fragment_create_promo_code, container, false);
 	}
 
 	@Override
-	public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+	public void onViewCreated(@NonNull final View view, @Nullable final Bundle savedInstanceState) {
 		super.onViewCreated(view, savedInstanceState);
 		_viewContainer = new ViewContainer(view);
 	}
@@ -73,31 +64,63 @@ public class CreatePromoCodeFragment extends BaseFragment {
 			return;
 		}
 		_creating = true;
-		final PromoCodeBuilder promoCodeBuilder = _viewContainer._promoCode.getText().length() == 0
-				? PromoCodeBuilder.createRandomCode()
-				: PromoCodeBuilder.createWithCode(_viewContainer._promoCode.getText().toString());
-		promoCodeBuilder.withMaxClaimCount(_viewContainer._maxClaim.getText().length() == 0 ? 0 : Integer.parseInt(_viewContainer._maxClaim.getText().toString()))
-				.withTimeLimit(_startTime, _endTime);
-		for (DynamicUi.DynamicInputHolder inputHolder : _customData) {
-			promoCodeBuilder.addData(inputHolder.getText(0), inputHolder.getText(1));
+		final PromoCodeContent promoCodeContent = _viewContainer._promoCode.getText().length() == 0
+						? PromoCodeContent.createRandomCode()
+						: PromoCodeContent.createWithCode(_viewContainer._promoCode.getText().toString());
+		promoCodeContent.withMaxClaimCount(_viewContainer._maxClaim.getText().length() == 0 ? 0 : Integer.parseInt(_viewContainer._maxClaim.getText().toString()))
+						.withTimeLimit(_startTime, _endTime);
+		for (final DynamicUi.DynamicInputHolder inputHolder : _customData) {
+			promoCodeContent.addProperty(inputHolder.getText(0), inputHolder.getText(1));
 		}
 
 		showLoading("Promo Code", "Creating...");
-		GetSocial.createPromoCode(promoCodeBuilder, new Callback<PromoCode>() {
-			@Override
-			public void onSuccess(PromoCode result) {
-				_creating = false;
-				hideLoading();
-				((PromoCodesFragment) getFragmentManager().findFragmentByTag("promocodes")).showFullPromoCodeDialog(result);
-			}
-
-			@Override
-			public void onFailure(GetSocialException exception) {
-				_creating = false;
-				hideLoading();
-				Toast.makeText(getContext(), "Failed to create promo code: " + exception.getMessage(), Toast.LENGTH_SHORT).show();
-			}
+		PromoCodes.create(promoCodeContent, result -> {
+			_creating = false;
+			hideLoading();
+			((PromoCodesFragment) getFragmentManager().findFragmentByTag("promocodes")).showFullPromoCodeDialog(result);
+		}, error -> {
+			_creating = false;
+			hideLoading();
+			Toast.makeText(getContext(), "Failed to create promo code: " + error.getMessage(), Toast.LENGTH_SHORT).show();
 		});
+	}
+
+	private void showDateTimePicker(final String title, final DateTimePickerCallback callback) {
+		final LinearLayout linearLayout = new LinearLayout(getContext());
+		linearLayout.setOrientation(LinearLayout.VERTICAL);
+		final DatePicker datePicker = new DatePicker(getContext());
+		final TimePicker timePicker = new TimePicker(getContext(), null, R.style.CustomDatePickerDialogTheme);
+
+		linearLayout.addView(datePicker, new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
+		linearLayout.addView(timePicker);
+
+		new AlertDialog.Builder(getContext())
+						.setTitle(title)
+						.setView(linearLayout)
+						.setPositiveButton("Set", new DialogInterface.OnClickListener() {
+							@Override
+							public void onClick(final DialogInterface dialog, final int which) {
+								dialog.dismiss();
+								final Calendar calendar = Calendar.getInstance();
+								calendar.set(
+												datePicker.getYear(), datePicker.getMonth(), datePicker.getDayOfMonth(),
+												timePicker.getCurrentHour(), timePicker.getCurrentMinute()
+								);
+								callback.onDatePicked(calendar.getTime());
+							}
+						})
+						.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
+							@Override
+							public void onClick(final DialogInterface dialog, final int which) {
+								dialog.dismiss();
+							}
+						})
+						.create()
+						.show();
+	}
+
+	private interface DateTimePickerCallback {
+		void onDatePicked(Date date);
 	}
 
 	public class ViewContainer {
@@ -117,7 +140,7 @@ public class CreatePromoCodeFragment extends BaseFragment {
 		@BindView(R.id.end_time)
 		TextView _endTimeLabel;
 
-		ViewContainer(View view) {
+		ViewContainer(final View view) {
 			ButterKnife.bind(this, view);
 		}
 
@@ -130,7 +153,7 @@ public class CreatePromoCodeFragment extends BaseFragment {
 		public void changeStartTime() {
 			showDateTimePicker("Select Start Time", new DateTimePickerCallback() {
 				@Override
-				public void onDatePicked(Date date) {
+				public void onDatePicked(final Date date) {
 					_startTime = date;
 					_startTimeLabel.setText("Start time: " + date.toString());
 				}
@@ -141,7 +164,7 @@ public class CreatePromoCodeFragment extends BaseFragment {
 		public void changeEndTime() {
 			showDateTimePicker("Select End Time", new DateTimePickerCallback() {
 				@Override
-				public void onDatePicked(Date date) {
+				public void onDatePicked(final Date date) {
 					_endTime = date;
 					_endTimeLabel.setText("End time: " + date.toString());
 				}
@@ -165,43 +188,5 @@ public class CreatePromoCodeFragment extends BaseFragment {
 			doCreatePromoCode();
 		}
 
-	}
-
-	private void showDateTimePicker(String title, final DateTimePickerCallback callback) {
-		final LinearLayout linearLayout = new LinearLayout(getContext());
-		linearLayout.setOrientation(LinearLayout.VERTICAL);
-		final DatePicker datePicker = new DatePicker(getContext());
-		final TimePicker timePicker = new TimePicker(getContext(), null, R.style.CustomDatePickerDialogTheme);
-
-		linearLayout.addView(datePicker, new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT));
-		linearLayout.addView(timePicker);
-
-		new AlertDialog.Builder(getContext())
-				.setTitle(title)
-				.setView(linearLayout)
-				.setPositiveButton("Set", new DialogInterface.OnClickListener() {
-					@Override
-					public void onClick(DialogInterface dialog, int which) {
-						dialog.dismiss();
-						Calendar calendar = Calendar.getInstance();
-						calendar.set(
-								datePicker.getYear(), datePicker.getMonth(), datePicker.getDayOfMonth(),
-								timePicker.getCurrentHour(), timePicker.getCurrentMinute()
-						);
-						callback.onDatePicked(calendar.getTime());
-					}
-				})
-				.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
-					@Override
-					public void onClick(DialogInterface dialog, int which) {
-						dialog.dismiss();
-					}
-				})
-				.create()
-				.show();
-	}
-
-	private interface DateTimePickerCallback {
-		void onDatePicked(Date date);
 	}
 }
